% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive.R
\name{adaptive_arrhythmic_melody_trials}
\alias{adaptive_arrhythmic_melody_trials}
\title{Adaptive arrhythmic melody trials block}
\usage{
adaptive_arrhythmic_melody_trials(
  label,
  num_items,
  item_bank,
  model,
  fixed_effects,
  demo = FALSE,
  page_type = "record_audio_page",
  page_title = "Sing back the melody",
  page_text = "Sing back the melody",
  get_answer = musicassessr::get_answer_pyin_melodic_production,
  give_first_melody_note = FALSE,
  play_melody_loop = FALSE,
  answer_column = "melody",
  feedback = FALSE
)
}
\arguments{
\item{feedback}{}
}
\value{

}
\description{
Adaptive arrhythmic melody trials block
}
