% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{add_feedback}
\alias{add_feedback}
\title{Helper for adding a feedback function to a timeline of pages}
\usage{
add_feedback(items, feedback, after = 2)
}
\arguments{
\item{after}{}
}
\value{

}
\description{
Helper for adding a feedback function to a timeline of pages
}
