% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{add_feedback_with_progress}
\alias{add_feedback_with_progress}
\title{Helper for adding a feedback function to a timeline of pages with a progress bar}
\usage{
add_feedback_with_progress(items, feedback, after = 2, scale_pr = 1)
}
\arguments{
\item{scale_pr}{}
}
\description{
Helper for adding a feedback function to a timeline of pages with a progress bar
}
