% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{arrhythmic_ntile_sampler_codeblock}
\alias{arrhythmic_ntile_sampler_codeblock}
\title{Sample arrhythmic melodies by N-tiles}
\usage{
arrhythmic_ntile_sampler_codeblock(
  item_bank,
  num_items,
  id = "arrhythmic_melody",
  phase = "test",
  n = 4,
  ntile_sampling_upper_bound = 4L
)
}
\arguments{
\item{n}{}
}
\description{
Sample arrhythmic melodies by N-tiles
}
