% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{audio_melodic_production_trials}
\alias{audio_melodic_production_trials}
\title{Present a trial block of melodies from audio files}
\usage{
audio_melodic_production_trials(
  audio_directory,
  no_to_sample = NULL,
  module_prefix = "audio_trial",
 
    text = "Click below to hear the melody. Sing back the melody. Click Stop when finished.",
  grab_meta_data
)
}
\arguments{
\item{grab_meta_data}{}
}
\value{

}
\description{
Present a trial block of melodies from audio files
}
