% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_builders.R
\name{build_multi_play_long_tone_record_audio_pages}
\alias{build_multi_play_long_tone_record_audio_pages}
\title{Build multiple play_long_tone_record_audio_pages based on a user's range}
\usage{
build_multi_play_long_tone_record_audio_pages(
  no_items,
  page_type = "record_audio_page",
  example = FALSE,
  feedback = FALSE
)
}
\arguments{
\item{top_range}{}
}
\value{

}
\description{
Build multiple play_long_tone_record_audio_pages based on a user's range
}
