% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy_long_note.R
\name{calculate_stable_part}
\alias{calculate_stable_part}
\title{Calculate metrics related to "changepoints" in a time series (e.g., for long note trials)
Input: df with "freq" column}
\usage{
calculate_stable_part(df, plot = FALSE, return_df = FALSE)
}
\arguments{
\item{return_df}{}
}
\description{
Calculate metrics related to "changepoints" in a time series (e.g., for long note trials)
Input: df with "freq" column
}
