% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_init.R
\name{call_api_on_start}
\alias{call_api_on_start}
\title{Call API on start}
\usage{
call_api_on_start(
  experiment_id = NA,
  experiment_condition_id = NA,
  user_id = NULL
)
}
\arguments{
\item{user_id}{}
}
\description{
Call API on start
}
