% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{classify_duration}
\alias{classify_duration}
\title{Classify ioi class (see Frieler.. )}
\usage{
classify_duration(dur_vec, ref_duration = 0.5)
}
\arguments{
\item{dur_vec}{best, should be a ioi vector}

\item{ref_duration}{}
}
\value{

}
\description{
Classify ioi class (see Frieler.. )
}
