% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_melodic_production.R
\name{compute_accuracy_measures_aligned}
\alias{compute_accuracy_measures_aligned}
\title{Compute aligned accuracy measures}
\usage{
compute_accuracy_measures_aligned(stimuli, recall)
}
\arguments{
\item{recall}{}
}
\description{
Compute aligned accuracy measures
}
