% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{conditional_proceed_if_yes}
\alias{conditional_proceed_if_yes}
\title{Conditional logic if posed a yes or no question}
\usage{
conditional_proceed_if_yes(question_page, logic_if_yes)
}
\arguments{
\item{question_page}{Should be an NAFC_page with Yes and No only options}

\item{logic_if_use}{Logic to execute if the user clicks Yes.}
}
\description{
Conditional logic if posed a yes or no question
}
