% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to a postgres DB}
\usage{
connect_to_db(
  host = Sys.getenv("DB_HOST"),
  db_name = Sys.getenv("DB_NAME"),
  username = Sys.getenv("DB_USER"),
  password = Sys.getenv("DB_PASSWORD")
)
}
\arguments{
\item{password}{}
}
\value{

}
\description{
Connect to a postgres DB
}
