% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_midi_or_audio_ui.R
\name{create_db_vars_template}
\alias{create_db_vars_template}
\title{Create a template for db_vars}
\usage{
create_db_vars_template(init_with_time_started = TRUE)
}
\arguments{
\item{init_with_time_started}{}
}
\description{
Create a template for db_vars
}
