% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{create_socials}
\alias{create_socials}
\title{Create an interface to share test score on socials}
\usage{
create_socials(socials, test_name, score, hashtag, test, url)
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Create an interface to share test score on socials
}
