% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_music.R
\name{default_range}
\alias{default_range}
\title{Get the default range for an instrument}
\usage{
default_range(instrument)
}
\arguments{
\item{instrument}{}
}
\description{
Get the default range for an instrument
}
