% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\title{Use musicassessr standard dictionary plus concatenate with your own}
\usage{
dict(additional_dict = NULL, main_dict = musicassessr::musicassessr_dict_df)
}
\arguments{
\item{main_dict}{}
}
\description{
Use musicassessr standard dictionary plus concatenate with your own
}
