% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_audio_from_ec2.R
\name{download_files_from_ec2}
\alias{download_files_from_ec2}
\title{Download files from the ec2 instance (in batches)}
\usage{
download_files_from_ec2(
  file_name,
  absolute_url,
  download_location,
  batch_size = 100
)
}
\arguments{
\item{batch_size}{}
}
\description{
Download files from the ec2 instance (in batches)
}
