% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{elt_add_final_session_info_to_db}
\alias{elt_add_final_session_info_to_db}
\title{Add final session info to postgres db}
\usage{
elt_add_final_session_info_to_db()
}
\description{
Add final session info to postgres db
}
