% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{elt_add_session_to_db}
\alias{elt_add_session_to_db}
\title{Add session to postgres db}
\usage{
elt_add_session_to_db(
  take_scores_from = c("db_table", "session_variable"),
  get_local = FALSE,
  test = NULL
)
}
\arguments{
\item{test}{}
}
\value{

}
\description{
Add session to postgres db
}
