% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{empty_page}
\alias{empty_page}
\title{An empty page (with trigger button hidden and possible to show on demand.)}
\usage{
empty_page(
  body = "",
  page_title = "",
  label = "",
  get_answer = function() { },
  save_answer = FALSE,
  button_text = "Next"
)
}
\arguments{
\item{button_text}{}
}
\value{

}
\description{
An empty page (with trigger button hidden and possible to show on demand.)
}
