% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy.R
\name{end_of_long_note_trial_screening}
\alias{end_of_long_note_trial_screening}
\title{End of long note trial screening code block}
\usage{
end_of_long_note_trial_screening()
}
\value{

}
\description{
End of long note trial screening code block
}
