% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_string_df_row}
\alias{expand_string_df_row}
\title{Expand a string dataframe row to a dataframe}
\usage{
expand_string_df_row(df, row_id = NULL)
}
\arguments{
\item{row_id}{}
}
\value{

}
\description{
Expand a string dataframe row to a dataframe
}
