% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_utils.R
\name{fake_range}
\alias{fake_range}
\title{Set a dummy range (to avoid having to manually set in the test)}
\usage{
fake_range(bottom_range = 60, top_range = 72)
}
\arguments{
\item{top_range}{}
}
\value{

}
\description{
Set a dummy range (to avoid having to manually set in the test)
}
