% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_image}
\alias{feedback_image}
\title{Feedback as an image}
\usage{
feedback_image(
  image,
  height = NULL,
  width = NULL,
  text = "Well done!",
  progress = NULL
)
}
\arguments{
\item{progress}{}
}
\description{
Feedback as an image
}
