% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_melodic_production}
\alias{feedback_melodic_production}
\title{Feedback for melodic production pages}
\usage{
feedback_melodic_production(melody_dtw = TRUE, answer_meta_data = TRUE)
}
\arguments{
\item{answer_meta_data}{}
}
\description{
Feedback for melodic production pages
}
