% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_melodic_production_simple}
\alias{feedback_melodic_production_simple}
\title{Simple melodic production feedback}
\usage{
feedback_melodic_production_simple()
}
\description{
Simple melodic production feedback
}
