% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_rhythm_production.R
\name{feedback_rhythm_production}
\alias{feedback_rhythm_production}
\title{Feedback for rhythm production style pages}
\usage{
feedback_rhythm_production()
}
\description{
Feedback for rhythm production style pages
}
