% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_syllable_classification}
\alias{feedback_syllable_classification}
\title{Feedback for syllable classification}
\usage{
feedback_syllable_classification()
}
\description{
Feedback for syllable classification
}
