% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{final_results}
\alias{final_results}
\title{Present Final Results}
\usage{
final_results(
  test_name,
  url,
  num_items_long_tone,
  num_items_arrhythmic,
  num_items_rhythmic
)
}
\arguments{
\item{num_items_rhythmic}{}
}
\value{

}
\description{
Present Final Results
}
