% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{final_results_pbet}
\alias{final_results_pbet}
\title{Present Final Results PBET}
\usage{
final_results_pbet(
  test_name = "PBET",
  url = "https://adaptiveeartraining.com",
  num_items_find_this_note = 0L,
  num_items_arrhythmic = 0L,
  num_items_rhythmic = 0L
)
}
\arguments{
\item{num_items_rhythmic}{}
}
\value{

}
\description{
Present Final Results PBET
}
