% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy_long_note.R
\name{find_closest_stimuli_pitch_to_user_production_pitches}
\alias{find_closest_stimuli_pitch_to_user_production_pitches}
\title{Find the closest pitch(s) in a stimulus, to the notes a user produced}
\usage{
find_closest_stimuli_pitch_to_user_production_pitches(
  stimuli_pitches,
  user_production_pitches,
  allOctaves = TRUE
)
}
\arguments{
\item{allOctaves}{}
}
\value{

}
\description{
Find the closest pitch(s) in a stimulus, to the notes a user produced
}
