% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{get_all_ngrams}
\alias{get_all_ngrams}
\title{Get all ngrams from a given vector}
\usage{
get_all_ngrams(x, N = 3)
}
\arguments{
\item{N}{}
}
\value{

}
\description{
Get all ngrams from a given vector
}
