% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_add_trial_and_compute_trial_scores_s3}
\alias{get_answer_add_trial_and_compute_trial_scores_s3}
\title{Get answer}
\usage{
get_answer_add_trial_and_compute_trial_scores_s3(
  input,
  state,
  midi_vs_audio = c("audio", "midi"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Get answer
}
