% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_async_midi_vs_audio}
\alias{get_answer_async_midi_vs_audio}
\title{Get answer async midi vs audio}
\usage{
get_answer_async_midi_vs_audio(input, state, ...)
}
\arguments{
\item{...}{}
}
\description{
Get answer async midi vs audio
}
