% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_midi_melodic_production}
\alias{get_answer_midi_melodic_production}
\title{Get a MIDI result and compute melodic production scores}
\usage{
get_answer_midi_melodic_production(input, state, ...)
}
\arguments{
\item{...}{}
}
\description{
Get a MIDI result and compute melodic production scores
}
