% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_null}
\alias{get_answer_null}
\title{Dummy get_answer function}
\usage{
get_answer_null(input, state, text = "example_page", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Dummy get_answer function
}
