% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_onset_detection}
\alias{get_answer_onset_detection}
\title{Get onset information from an audio file}
\usage{
get_answer_onset_detection(
  input,
  state,
  add_silence_to_beginning_of_audio_file = 0.5,
  add_silence_to_end_of_audio_file = 0,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Get onset information from an audio file
}
