% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin}
\alias{get_answer_pyin}
\title{Get answer by running pyin on a recorded audio file}
\usage{
get_answer_pyin(
  input,
  type = c("both", "notes", "pitch_track"),
  state,
  melconv = FALSE,
  add_silence_to_beginning_of_audio_file = 0.5,
  add_silence_to_end_of_audio_file = 0,
  ...
)
}
\arguments{
\item{add_silence_to_beginning_of_audio_file}{In seconds.}

\item{add_silence_to_end_of_audio_file}{In seconds.}

\item{...}{}
}
\description{
Get answer by running pyin on a recorded audio file
}
