% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin_long_note}
\alias{get_answer_pyin_long_note}
\title{get_answer pyin for long notes}
\usage{
get_answer_pyin_long_note(input, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
get_answer pyin for long notes
}
