% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin_mel_prod_only}
\alias{get_answer_pyin_mel_prod_only}
\title{Get mel production scores, but without singing/pitch accuracy measures}
\usage{
get_answer_pyin_mel_prod_only(
  input,
  type = c("both", "note", "pitch_track"),
  state,
  melconv = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Get mel production scores, but without singing/pitch accuracy measures
}
