% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin_melodic_production}
\alias{get_answer_pyin_melodic_production}
\title{Analyse a melody with pyin then compute melodic production scores}
\usage{
get_answer_pyin_melodic_production(
  input,
  type = c("both", "notes", "pitch_track"),
  state,
  melconv = FALSE,
  additional_scoring_measures = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Analyse a melody with pyin then compute melodic production scores
}
