% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin_melodic_production_additional_measures}
\alias{get_answer_pyin_melodic_production_additional_measures}
\title{Wrapper to add user-specified additional scoring measures to pYIN melodic production}
\usage{
get_answer_pyin_melodic_production_additional_measures(
  type = c("both", "notes", "pitch_track"),
  melconv = FALSE,
  additional_scoring_measures = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Wrapper to add user-specified additional scoring measures to pYIN melodic production
}
