% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_rhythm_production}
\alias{get_answer_rhythm_production}
\title{Get answer for rhythm production data}
\usage{
get_answer_rhythm_production(
  input,
  state,
  type = c("midi", "audio", "key_presses"),
  feedback = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Get answer for rhythm production data
}
