% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_save_audio_file}
\alias{get_answer_save_audio_file}
\title{Simple saving of the corresponding URL of a trial file}
\usage{
get_answer_save_audio_file(input, ...)
}
\arguments{
\item{...}{}
}
\description{
Simple saving of the corresponding URL of a trial file
}
