% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_save_aws_key}
\alias{get_answer_save_aws_key}
\title{Simple saving of the corresponding URL of a trial file}
\usage{
get_answer_save_aws_key(input, ...)
}
\arguments{
\item{...}{}
}
\description{
Simple saving of the corresponding URL of a trial file
}
