% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_simple_pyin_summary}
\alias{get_answer_simple_pyin_summary}
\title{summarise pyin output with basic statistics}
\usage{
get_answer_simple_pyin_summary(input, state, ...)
}
\arguments{
\item{...}{}
}
\description{
summarise pyin output with basic statistics
}
