% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_store_async}
\alias{get_answer_store_async}
\title{Get answer and recode the last answer if its corresponding promise has resolved}
\usage{
get_answer_store_async(input, state, opt, scoring = "melody", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Get answer and recode the last answer if its corresponding promise has resolved
}
