% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_syllable_classification}
\alias{get_answer_syllable_classification}
\title{Estimate syllable from an audio file}
\usage{
get_answer_syllable_classification(input, state, ...)
}
\arguments{
\item{state}{}
}
\description{
Estimate syllable from an audio file
}
