% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_wav_then_pyin}
\alias{get_answer_wav_then_pyin}
\title{Produce wav from audio recorded in browser and return pyin result}
\usage{
get_answer_wav_then_pyin(input, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Produce wav from audio recorded in browser and return pyin result
}
