% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_range.R
\name{get_audio_file_job_process}
\alias{get_audio_file_job_process}
\title{Get audio file job process}
\usage{
get_audio_file_job_process(
  asynchronous_api_mode,
  state_var_name = "wav_file",
  var_name_to_set_result,
  var_to_take_from_message = NULL
)
}
\arguments{
\item{var_to_take_from_message}{}
}
\description{
Get audio file job process
}
