% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_range.R
\name{get_instrument_range_pages}
\alias{get_instrument_range_pages}
\title{Get Instrument Range Pages}
\usage{
get_instrument_range_pages(
  input_type,
  show_musical_notation = FALSE,
  adjust_range = FALSE,
  test_type = c("voice", "instrument"),
  concise_wording = FALSE,
  asynchronous_api_mode = FALSE
)
}
\arguments{
\item{asynchronous_api_mode}{}
}
\description{
Get Instrument Range Pages
}
