% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy_long_note.R
\name{get_long_note_pcas}
\alias{get_long_note_pcas}
\title{Get long note PCA scores}
\usage{
get_long_note_pcas(long_note_scores)
}
\arguments{
\item{long_note_beginning_of_second_cpt}{}
}
\value{

}
\description{
Get long note PCA scores
}
