% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_longitudinal_session_items}
\alias{get_longitudinal_session_items}
\title{Get longitudinal study items}
\usage{
get_longitudinal_session_items(participant_number, num_test_session)
}
\arguments{
\item{num_test_session}{}
}
\description{
Get longitudinal study items
}
