% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_melody_dtw.R
\name{get_melody_dtw}
\alias{get_melody_dtw}
\title{Compute the dynamic time warp distance between a melodic stimuli and a user's continuous pitch curve, attempting to sing that stimuli}
\usage{
get_melody_dtw(stimuli, stimuli_durations, pyin_pitch_track, pyin_res)
}
\arguments{
\item{pyin_res}{}
}
\value{

}
\description{
Compute the dynamic time warp distance between a melodic stimuli and a user's continuous pitch curve, attempting to sing that stimuli
}
