% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy.R
\name{get_note_precision}
\alias{get_note_precision}
\title{Get note precision measure}
\usage{
get_note_precision(pyin_result)
}
\arguments{
\item{pyin_result}{}
}
\value{

}
\description{
Get note precision measure
}
