% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{get_review_trials}
\alias{get_review_trials}
\title{Get a random selection of previous trials based on a user ID}
\usage{
get_review_trials(no_reviews, state)
}
\arguments{
\item{no_reviews}{}
}
\description{
Note currently this doesn't distinguish between the most recent session and others further back, nor the trial type previously (arrhythmic, rhythmic etc.)
}
