% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR.R
\name{get_SNR_pages}
\alias{get_SNR_pages}
\title{Get SNR pages}
\usage{
get_SNR_pages(
  min_SNR = 14,
  absolute_url = character(),
  report_SNR = FALSE,
  allow_SNR_failure = FALSE
)
}
\arguments{
\item{allow_SNR_failure}{}
}
\description{
Get SNR pages
}
