% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR.R
\name{get_SNR_pages_loop}
\alias{get_SNR_pages_loop}
\title{Get SNR pages as a loop (i.e., the participant can take multiple attempts)}
\usage{
get_SNR_pages_loop(
  min_SNR = 14,
  absolute_url = character(),
  report_SNR,
  allow_SNR_failure = FALSE,
  asynchronous_api_mode = FALSE
)
}
\arguments{
\item{asynchronous_api_mode}{}
}
\description{
Get SNR pages as a loop (i.e., the participant can take multiple attempts)
}
