% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_AWS.R
\name{get_table}
\alias{get_table}
\title{Get a postgres table}
\usage{
get_table(name = c("trials", "sessions", "production"))
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Get a postgres table
}
