% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_deprecated.R
\name{get_user_info_page}
\alias{get_user_info_page}
\title{get_user_info_page}
\arguments{
\item{chrome_only}{Logical. If \code{TRUE}, only Chrome browser is recommended. Default is \code{TRUE}.}
}
\value{
None.
}
\description{
Check a user has the requirements to run our test and record their browser information.
}
\details{
This function is deprecated and will be removed in a future version.
}
