% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singing_range.R
\name{get_voice_range_page}
\alias{get_voice_range_page}
\title{A page to select a voice range}
\usage{
get_voice_range_page(with_examples = TRUE)
}
\arguments{
\item{with_examples}{}
}
\description{
A page to select a voice range
}
