% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{grab_meta_data}
\alias{grab_meta_data}
\title{Grab meta data from a df}
\usage{
grab_meta_data(meta_data_df, lookup_column, value)
}
\arguments{
\item{value}{}
}
\value{

}
\description{
Grab meta data from a df
}
