% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{harmcore}
\alias{harmcore}
\title{Compute harmcore}
\usage{
harmcore(
  pitch_vec1,
  pitch_vec2,
  segmentation1 = NULL,
  segmentation2 = NULL,
  segmentation_type = c("phrase_boundary_marker", "segment_id")
)
}
\arguments{
\item{segmentation_type}{}
}
\description{
Compute harmcore
}
