% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_requirements}
\alias{have_requirements}
\title{Check if user has requirements for musicassessr test}
\usage{
have_requirements(answer, ...)
}
\arguments{
\item{...}{}
}
\description{
Check if user has requirements for musicassessr test
}
