% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_item_into_every_other_n_position_in_list}
\alias{insert_item_into_every_other_n_position_in_list}
\title{Insert item into every other n position in list}
\usage{
insert_item_into_every_other_n_position_in_list(l, item_to_add, n = 2)
}
\arguments{
\item{n}{}
}
\description{
Insert item into every other n position in list
}
