% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{interval_perception_trials}
\alias{interval_perception_trials}
\title{A block which test a participant's perception of intervals}
\usage{
interval_perception_trials(
  num_items = 26L,
  sound = "piano",
  page_title = psychTestR::i18n("interval_perception_page_title"),
  instruction_text =
    shiny::tags$div(shiny::tags$p(psychTestR::i18n("interval_perception_instruction_1")),
    shiny::tags$p(psychTestR::i18n("interval_perception_instruction_2")),
    shiny::tags$p(psychTestR::i18n("interval_perception_instruction_3")),
    shiny::tags$p(shiny::tags$em(psychTestR::i18n("interval_perception_instruction_4"))))
)
}
\arguments{
\item{instruction_text}{}
}
\description{
A block which test a participant's perception of intervals
}
