% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{item_sampler_materialized_view}
\alias{item_sampler_materialized_view}
\title{Item sampler materialized view}
\usage{
item_sampler_materialized_view(
  db_con,
  no_items,
  table = "Berkowitz_ngram_n_view",
  shuffle = TRUE
)
}
\arguments{
\item{shuffle}{}
}
\description{
Item sampler materialized view
}
