% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{item_sampler_v1}
\alias{item_sampler_v1}
\title{Item sampler (stratified sampling)}
\usage{
item_sampler_v1(item_bank, no_items, replace = FALSE, shuffle = TRUE)
}
\arguments{
\item{shuffle}{}
}
\description{
Item sampler (stratified sampling)
}
