% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{item_sampler_v2}
\alias{item_sampler_v2}
\title{Item sampler (stratified sampling) v2}
\usage{
item_sampler_v2(item_bank, no_items, replace = FALSE, shuffle = TRUE)
}
\arguments{
\item{shuffle}{}
}
\description{
Item sampler (stratified sampling) v2
}
