% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\docType{data}
\name{long_note_agg}
\alias{long_note_agg}
\title{Aggregated long note data from SAA paper}
\format{
Dataframe.
}
\usage{
long_note_agg
}
\description{
Aggregated long note data from SAA paper
}
\keyword{datasets}
