% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy.R
\name{long_note_pitch_metrics}
\alias{long_note_pitch_metrics}
\title{Compute long tone pitch metrics}
\usage{
long_note_pitch_metrics(target_pitch, freq)
}
\arguments{
\item{freq}{}
}
\value{

}
\description{
Compute long tone pitch metrics
}
