% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS.R
\name{make_aws_credentials_global}
\alias{make_aws_credentials_global}
\title{Make AWS Credentials Global}
\usage{
make_aws_credentials_global(aws_credentials)
}
\arguments{
\item{aws_credentials}{}
}
\value{

}
\description{
Make AWS Credentials Global
}
