% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21.R
\name{MAST21_trials}
\alias{MAST21_trials}
\title{Deploy a block of the MAST21 stimuli}
\usage{
MAST21_trials(
  item_bank,
  num_items,
  num_examples = NULL,
  feedback = FALSE,
  get_answer = musicassessr::get_answer_pyin,
  sound = "piano",
  page_text = psychTestR::i18n("sing_melody_trial"),
  page_title_melody = "Please sing the melody",
  page_title_long_note = "Sing the note with a \\"Dooo\\" sound.",
  page_text_long_note = "Please sing the note after you hear it, then click Stop.",
 
    instruction_text = "Now you will hear melodies with rhythms. Please try and sing the melodies with the correct rhythm."
)
}
\arguments{
\item{microphone_calibration_page}{}
}
\value{

}
\description{
Deploy a block of the MAST21 stimuli
}
