% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melconv.R
\name{melconv_from_pyin_res}
\alias{melconv_from_pyin_res}
\title{#' Melconv
#'
#' @param file_name
#' @param return_notes_and_durs
#'
#' @return
#' @export
#'
#' @examples
melconv <- function(file_name, return_notes_and_durs = TRUE) {}
\usage{
melconv_from_pyin_res(pyin_res)
}
\arguments{
\item{pyin_res}{}
}
\value{

}
\description{
#' Melconv
#'
#' @param file_name
#' @param return_notes_and_durs
#'
#' @return
#' @export
#'
#' @examples
melconv <- function(file_name, return_notes_and_durs = TRUE) {
}
\details{

}
\section{then use melconv}{
melconv_res <- system2(command = 'melconv',
args = c('-f midi',
paste0('-i ', file_name),
'-o /srv/shiny-server/files/mid/'),
stdout = TRUE, stderr = FALSE)
}

