% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{melody_scoring_from_user_input}
\alias{melody_scoring_from_user_input}
\title{Score melodic production measures}
\usage{
melody_scoring_from_user_input(
  input,
  result = NULL,
  trial_type = "audio",
  user_melody_input = NULL,
  singing_measures = TRUE,
  pyin_pitch_track = NULL,
  stimuli = NA,
  stimuli_durations = NA
)
}
\arguments{
\item{stimuli_durations}{}
}
\value{

}
\description{
Score melodic production measures
}
