% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody_str_to_vector.R
\name{melody_str_to_vector}
\alias{melody_str_to_vector}
\title{String melody to vector}
\usage{
melody_str_to_vector(mel_str, sep)
}
\arguments{
\item{mel_str}{character}

\item{sep}{character}
}
\value{
integer
}
\description{
Takes a melody, written as a string of numbers (e.g MIDI) notes separated by a character (e.g a comma) and returns it as a numeric vector
}
\examples{
melody_str_to_vector(mel_str = "60, 62, 64, 65", sep = ",")
}
