% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microphone_calibration_page.R
\name{microphone_calibration_page}
\alias{microphone_calibration_page}
\title{Microphone Calibration Page}
\usage{
microphone_calibration_page(
  admin_ui = NULL,
  on_complete = NULL,
  label = NULL,
 
    body = shiny::tags$div(shiny::tags$p(psychTestR::i18n("microphone_calibration_message_1")),
    shiny::tags$p(psychTestR::i18n("microphone_calibration_message_2")),
    shiny::tags$p(psychTestR::i18n("microphone_calibration_message_3"))),
  button_text = psychTestR::i18n("microphone_calibration_button"),
  save_answer = FALSE,
  deploy_crepe_stats = FALSE
)
}
\arguments{
\item{deploy_crepe_stats}{}
}
\value{

}
\description{
Microphone Calibration Page
}
