% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midi_or_audio_reactive}
\alias{midi_or_audio_reactive}
\title{A conditional to present either MIDI OR audio logic, depending on test time user decision}
\usage{
midi_or_audio_reactive(midi_logic, audio_logic)
}
\arguments{
\item{audio_logic}{}
}
\description{
A conditional to present either MIDI OR audio logic, depending on test time user decision
}
