% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{multi_interval_page}
\alias{multi_interval_page}
\title{Deploy a block of n trials which play back intervals}
\usage{
multi_interval_page(
  num_items = 26L,
  page_title = psychTestR::i18n("interval_perception_page_title")
)
}
\arguments{
\item{page_title}{}
}
\description{
Deploy a block of n trials which play back intervals
}
