% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_melody_loop.R
\name{multi_page_play_melody_loop}
\alias{multi_page_play_melody_loop}
\title{A page builder for creating a specified number of play_melody_loops}
\usage{
multi_page_play_melody_loop(
  item_bank = NULL,
  presampled_items = NULL,
  num_items = NULL,
  var_name = "melody",
  stimuli_type = "midi_notes",
  page_type = "record_audio_page",
  max_goes = 4L,
  page_title = psychTestR::i18n("copy_melody_title"),
  page_text =
    "Press play to hear the melody, then play it back as best as you can when it finishes.",
  get_answer = get_answer_pyin_melodic_production,
  rel_to_abs_mel_function = NULL,
  start_from_trial_no = 1L,
  clip_stimuli_length = FALSE,
  arrhythmic = FALSE,
  example = FALSE,
  feedback = FALSE,
  sound = "piano",
  get_trial_characteristics_function = NULL,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_progress = TRUE,
  sheet_music_start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  get_similarity_to_previous_melody = FALSE,
  volume_meter = FALSE,
  volume_meter_type = "default",
  melody_block_paradigm = c("standard", "sing_melody_first",
    "learn_phase_visual_display_modality"),
  singing_trials = TRUE,
  phase = c("test", "learn", "review", "example"),
  melody_trial_paradigm = c("call_and_response", "simultaneous_recall"),
  first_note_message = psychTestR::i18n("first_note_is"),
  transposed_message = psychTestR::i18n("transposed"),
  play_first_note_button_text = psychTestR::i18n("play_first_note"),
  learn_test_paradigm = FALSE,
  sample_item_bank_via_api = FALSE,
  pass_items_through_url_parameter = FALSE
)
}
\arguments{
\item{pass_items_through_url_parameter}{}
}
\description{
A page builder for creating a specified number of play_melody_loops
}
