% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_showcase.R
\name{musical_test_showcase}
\alias{musical_test_showcase}
\title{Launch the musical test showcase of musicassessr tests}
\usage{
musical_test_showcase(
  app_name = "test_showcase",
  musicassessr_aws = FALSE,
  feedback = TRUE,
  num_items_rtt = list(free_recall = 2L, sync_beat = 2L, call_and_response = 2L),
  num_items_saa = list(long_tones = 2L, arrhythmic = 2L, rhythmic = 2L),
  num_items_pbet = list(find_this_note = 0L, arrhythmic = list(key_hard = 2L, key_easy =
    2L), rhythmic = list(key_hard = 2L, key_easy = 2L), interval_perception = 0L,
    wjd_audio = list(key_easy = 0L, key_hard = 0L)),
  showcase_name = "musicassessr"
)
}
\arguments{
\item{showcase_name}{}
}
\description{
Launch the musical test showcase of musicassessr tests
}
