% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\docType{data}
\name{musicassessr_dict}
\alias{musicassessr_dict}
\title{The musicassessr dictionary}
\format{
psychTestR::i18n_dict
}
\usage{
musicassessr_dict
}
\description{
The musicassessr dictionary
}
\keyword{datasets}
