% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_js_scripts.R
\name{musicassessr_js}
\alias{musicassessr_js}
\title{musicassessr scripts}
\usage{
musicassessr_js(
  state = "production",
  visual_notation = FALSE,
  midi_file_playback = FALSE
)
}
\arguments{
\item{midi_file_playback}{}
}
\value{

}
\description{
musicassessr scripts
}
