% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_js_scripts.R
\name{musicassessr_js_scripts}
\alias{musicassessr_js_scripts}
\title{Import the required musicassessr dependencies into a webpage}
\usage{
musicassessr_js_scripts(musicassessr_state = NULL)
}
\arguments{
\item{musicassessr_state}{}
}
\value{

}
\description{
Import the required musicassessr dependencies into a webpage
}
