% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_musicassessr_test.R
\name{musicassessr_opt}
\alias{musicassessr_opt}
\title{Specify options for a musicassessr test.}
\usage{
musicassessr_opt(
  setup_pages = TRUE,
  setup_options = setup_pages_options(),
  app_name = "",
  midi_file_playback = FALSE,
  visual_notation = FALSE,
  record_audio = TRUE,
  midi_input = FALSE,
  experiment_id = NULL,
  experiment_condition_id = NULL,
  user_id = NULL,
  instrument_id = NULL,
  get_p_id = FALSE,
  asynchronous_api_mode = FALSE,
  default_range = set_default_range("Piano"),
  css = c("https://musicassessr.com/assets/css/style_songbird.css",
    system.file("www/css/musicassessr.css", package = "musicassessr")),
  username = NULL,
  get_pid_prompt = shiny::tags$div(shiny::tags$h3("Choose an anonymous ID"),
    shiny::tags$p("Use the first two letters of your first name, plus, in numbers, your month and year of birth."),
    shiny::tags$p("For example, Mike, born in December 1900, would be ",
    shiny::tags$em("mi121900"), ".")),
  instrument = "Voice",
  get_user_info = TRUE,
  redirect_on_failure_url = "https://www.google.com/",
  async_success_msg = if (is.scalar.character(username))
    paste0(psychTestR::i18n("Hello"), " ", username, "!") else
    psychTestR::i18n("lets_proceed"),
  use_presigned_url = TRUE
)
}
\arguments{
\item{setup_pages}{Should there be setup pages?}

\item{setup_options}{Setup page options.}

\item{app_name}{App name for audio recording apps.}

\item{midi_file_playback}{Will MIDI files need to be played back?}

\item{visual_notation}{Will there be visual notation?}

\item{record_audio}{Is audio recording required?}

\item{midi_input}{Is MIDI input required?}

\item{experiment_id}{An experiment ID.}

\item{experiment_condition_id}{An experiment condition ID.}

\item{user_id}{A user ID.}

\item{instrument_id}{An instrument ID.}

\item{get_p_id}{Should a participant ID get collected at the beginning of the test?}

\item{asynchronous_api_mode}{Should async API mode be on?}

\item{default_range}{Should there be default range set?}

\item{css}{A css stylesheet to use (passes to psychTestR).}

\item{username}{Hardcode a username.}

\item{get_pid_prompt}{What prompt to you want to use for the get_p_id page?}

\item{instrument}{What instrument is the test using?}

\item{get_user_info}{Get user geolocation and device information?}

\item{redirect_on_failure_url}{If the test fails, where should the participant be redirected?}

\item{async_success_msg}{What message should be shown after the user logs in via the async system?}

\item{use_presigned_url}{For audio uploading, should a presigned URL be used?}
}
\description{
Specify options for a musicassessr test.
}
