% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{ngrukkon}
\alias{ngrukkon}
\title{ngrukkon as in  Müllensiefen & Frieler (2004)}
\usage{
ngrukkon(x, y, N = 3)
}
\arguments{
\item{N}{}
}
\description{
ngrukkon as in  Müllensiefen & Frieler (2004)
}
