% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{ngrukkon_safe}
\alias{ngrukkon_safe}
\title{An ngrukkon wrapper to produce warnings and return NAs rather than stop if one entry is too short}
\usage{
ngrukkon_safe(x, y, N = 3)
}
\arguments{
\item{N}{}
}
\value{

}
\description{
An ngrukkon wrapper to produce warnings and return NAs rather than stop if one entry is too short
}
