% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{on_complete_recode_async_penultimate}
\alias{on_complete_recode_async_penultimate}
\title{Recode an asynchronous result through a psychTestR on_complete function}
\usage{
on_complete_recode_async_penultimate(input, state, opt, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Recode an asynchronous result through a psychTestR on_complete function
}
