% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{opti3}
\alias{opti3}
\title{Score using the opti3 measure of similarity}
\usage{
opti3(
  pitch_vec1,
  dur_vec1,
  pitch_vec2,
  dur_vec2,
  N = 3,
  use_bootstrap = TRUE,
  return_components = FALSE
)
}
\arguments{
\item{return_components}{}
}
\value{

}
\description{
Score using the opti3 measure of similarity
}
