% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{opti3_df}
\alias{opti3_df}
\title{opti3 for melodies read by read_melody
returns sorted tibble of transpositions of melody2 and opti3 similarity}
\usage{
opti3_df(melody1, melody2, N = 3, use_bootstrap = FALSE, return_winner = TRUE)
}
\arguments{
\item{return_winner}{}
}
\description{
opti3 for melodies read by read_melody
returns sorted tibble of transpositions of melody2 and opti3 similarity
}
