% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbet_learn_test.R
\name{pbet_learn_test_paradigm}
\alias{pbet_learn_test_paradigm}
\title{PBET learn test paradigm}
\usage{
pbet_learn_test_paradigm(
  no_trials = 4L,
  bpm = 100,
  mute_midi_playback = TRUE,
  asynchronous_api_mode = TRUE
)
}
\arguments{
\item{asynchronous_api_mode}{}
}
\description{
PBET learn test paradigm
}
