% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbet_learn_test.R
\name{pbet_learn_test_paradigm_standalone}
\alias{pbet_learn_test_paradigm_standalone}
\title{PBET learn test block}
\usage{
pbet_learn_test_paradigm_standalone(no_trials, asynchronous_api_mode = TRUE)
}
\arguments{
\item{asynchronous_api_mode}{}
}
\description{
PBET learn test block
}
