% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_long_tone_record_audio.R
\name{play_long_tone_record_audio_page}
\alias{play_long_tone_record_audio_page}
\title{Create a page which produces a given tone}
\usage{
play_long_tone_record_audio_page(
  note = NULL,
  long_note_no = NULL,
  long_tone_length = 5,
  page_title = psychTestR::i18n("long_tone_title"),
  page_text = psychTestR::i18n("long_tone_text"),
  play_button_text = psychTestR::i18n("Play"),
  page_type = "record_audio_page",
  example = FALSE,
  get_answer = get_answer_pyin_long_note,
  page_label = "long_tone",
  total_no_long_notes = 0,
  show_progress = FALSE,
  trial_paradigm = c("simultaneous_recall", "call_and_response"),
  call_and_response_end = c("manual", "auto"),
  singing_trial = TRUE,
  on_complete = NULL
)
}
\arguments{
\item{on_complete}{}
}
\description{
Create a page which produces a given tone
}
