% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_melody_loop.R
\name{play_melody_loop}
\alias{play_melody_loop}
\title{Create a psychTestR test loop for having several attempts at playing back a melody.}
\usage{
play_melody_loop(
  melody = NULL,
  melody_no = "x",
  var_name = "melody",
  stimuli_type = "midi_notes",
  max_goes = 3L,
  max_goes_forced = FALSE,
  page_type = "record_audio_page",
  page_title = "Copy The Melody",
 
    page_text = "Press play to hear the melody, then play it back as best as you can when it finishes.",
  answer_meta_data = data.frame(),
  get_answer = get_answer_pyin_melodic_production,
  rel_to_abs_mel_function = NULL,
  clip_stimuli_length = FALSE,
  start_note = 1L,
  end_note = "end",
  durations = "null",
  arrhythmic = FALSE,
  note_length = 0.5,
  play_button_text = psychTestR::i18n("Play"),
  example = FALSE,
  sound = "piano",
  reactive_stimuli = NULL,
  get_trial_characteristics_function = NULL,
  give_first_melody_note = FALSE
)
}
\arguments{
\item{give_first_melody_note}{}
}
\value{

}
\description{
Create a psychTestR test loop for having several attempts at playing back a melody.
}
