% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_melody_dtw.R
\name{plot_dtw_melody}
\alias{plot_dtw_melody}
\title{Plot a melody after being transformed into a representation for use with the dynamic time warp algorithm}
\usage{
plot_dtw_melody(stimuli, stimuli_durations, pyin_smoothed_pitchtrack)
}
\arguments{
\item{pyin_smoothed_pitchtrack}{}
}
\description{
Plot a melody after being transformed into a representation for use with the dynamic time warp algorithm
}
