% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_longitudinal.R
\name{plot_longitudinal}
\alias{plot_longitudinal}
\title{Plot longitudinal}
\usage{
plot_longitudinal(
  type = c("ability_across_sessions", "mean_opti3_across_sessions",
    "melody_by_session_average", "melody_by_attempt_and_sessions",
    "average_melody_opti3_across_sessions_for_repeated_melodies")
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Plot longitudinal
}
