% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_melody_dtw.R
\name{prepare_mel_stimuli_for_dtw}
\alias{prepare_mel_stimuli_for_dtw}
\title{Convert a melody to a pseudo-timeseries representation}
\usage{
prepare_mel_stimuli_for_dtw(melody, durations, convert_midi_to_freq = TRUE)
}
\arguments{
\item{convert_midi_to_freq}{}
}
\description{
Convert a melody to a pseudo-timeseries representation
}
