% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_music.R
\name{present_stimuli_midi_notes_both}
\alias{present_stimuli_midi_notes_both}
\title{Present midi notes in both visual and auditory modalities}
\usage{
present_stimuli_midi_notes_both(
  stimuli,
  note_length = 0.5,
  sound = "piano",
  asChord = FALSE,
  play_button_text = "Play",
  ascending = TRUE,
  visual_music_notation_id = "sheet_music",
  play_button_id = "playButton",
  button_area_id = "button_area",
  start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  show_sheet_music = FALSE,
  sheet_music_id = "sheet_music",
  page_type = "null",
  durations = NULL,
  trigger_end_of_stimuli_fun = "function() { console.log('Stimulus finished!'); }"
)
}
\arguments{
\item{...}{}
}
\description{
Present midi notes in both visual and auditory modalities
}
