% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_music.R
\name{present_stimuli_midi_notes_visual}
\alias{present_stimuli_midi_notes_visual}
\title{Present MIDI notes as musical notation}
\usage{
present_stimuli_midi_notes_visual(
  stimuli,
  note_length,
  asChord = FALSE,
  ascending,
  id = "sheet_music",
  present_div = TRUE,
  clef = "auto",
  give_first_melody_note = FALSE,
  transpose_first_melody_note = 0,
  show_first_melody_note_visual = FALSE,
  audio_play_button_id = "playButton"
)
}
\arguments{
\item{audio_play_button_id}{}
}
\value{

}
\description{
Present MIDI notes as musical notation
}
