% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{produce_extra_melodic_features}
\alias{produce_extra_melodic_features}
\title{Produce extra melodic features from a pyin note track}
\usage{
produce_extra_melodic_features(pyin_style_res)
}
\arguments{
\item{pyin_style_res}{}
}
\value{

}
\description{
Produce extra melodic features from a pyin note track
}
