% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production_perception_reflection_paradigm.R
\name{production_reflection_perception_paradigm_standalone}
\alias{production_reflection_perception_paradigm_standalone}
\title{Production reflection perception paradigm test}
\usage{
production_reflection_perception_paradigm_standalone(
  num_items_arrhythmic = 4L,
  num_items_rhythmic = 4L,
  max_attempts = 3L
)
}
\arguments{
\item{max_attempts}{}
}
\description{
Production reflection perception paradigm test
}
