% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-promises-pipe.R
\name{\%...>\%}
\alias{\%...>\%}
\title{Promise pipe operator
#'}
\usage{
lhs \%...>\% rhs
}
\arguments{
\item{lhs}{A promise object.}

\item{rhs}{A function which handles the result of the promise}
}
\value{
The result of the promise when it resolves.
}
\description{
Promise pipe operator
#'
}
\keyword{internal}
