% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psych_test_snap}
\alias{psych_test_snap}
\title{Useful function for randomising which of 2 blocks goes first}
\usage{
psych_test_snap(block1, block2)
}
\arguments{
\item{block2}{}
}
\value{

}
\description{
Useful function for randomising which of 2 blocks goes first
}
