% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{read_melody}
\alias{read_melody}
\title{read a pYIN note track (outputted from Sonic Annotator or Tony) and make it nice}
\usage{
read_melody(fname, style = c("sonic_annotator", "tony"))
}
\arguments{
\item{style}{}
}
\description{
read a pYIN note track (outputted from Sonic Annotator or Tony) and make it nice
}
