% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{record_and_present_audio_block}
\alias{record_and_present_audio_block}
\title{A block for playing melodies as audio then recording the participant response}
\usage{
record_and_present_audio_block(
  files_list,
  page_text = "Press play to hear a melody, then sing it back.",
  get_answer = musicassessr::get_answer_pyin
)
}
\arguments{
\item{get_answer}{}
}
\description{
A block for playing melodies as audio then recording the participant response
}
