% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_audio_page.R
\name{record_audio_page}
\alias{record_audio_page}
\title{Record Audio Page}
\usage{
record_audio_page(
  body = " ",
  label = "record_audio_page",
  stimuli = " ",
  stimuli_reactive = FALSE,
  page_text = " ",
  page_title = " ",
  interactive = FALSE,
  show_record_button = TRUE,
  get_answer = get_answer_store_async,
  answer_meta_data = data.frame(),
  show_aws_controls = FALSE,
  button_text = "Record",
  stop_button_text = "Stop",
  record_duration = NULL,
  on_complete = NULL,
  auto_next_page = FALSE,
  save_answer = TRUE,
  page_text_first = TRUE,
  happy_with_response = FALSE,
  attempts_left = integer(),
  max_goes_forced = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Record Audio Page
}
