% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{record_key_presses_block}
\alias{record_key_presses_block}
\title{A block of record key press pages}
\usage{
record_key_presses_block(
  no_pages,
  label = "record_key_presses",
  feedback = NULL,
  get_answer = get_answer_pyin,
  page_title = psychTestR::i18n("Record_audio"),
  page_text = psychTestR::i18n("click_to_record_audio")
)
}
\arguments{
\item{page_title}{}
}
\description{
A block of record key press pages
}
