% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{record_midi_block}
\alias{record_midi_block}
\title{A block of record MIDI pages}
\usage{
record_midi_block(
  no_pages,
  label = "record_midi_page",
  feedback = NULL,
  get_answer = get_answer_midi,
  page_title = psychTestR::i18n("Record_audio"),
  page_text = psychTestR::i18n("click_to_record_audio"),
  autoInstantiate = TRUE
)
}
\arguments{
\item{autoInstantiate}{}
}
\description{
A block of record MIDI pages
}
