% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{redirect_page}
\alias{redirect_page}
\title{A page for redirecting a participant after x ms}
\usage{
redirect_page(
  text = "Thank you, you will now be redirected.",
  ms = 5000,
  url = "http://www.google.com",
  final = TRUE
)
}
\arguments{
\item{final}{}
}
\description{
A page for redirecting a participant after x ms
}
