% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition.R
\name{rel_to_abs_mel_mean_centred}
\alias{rel_to_abs_mel_mean_centred}
\title{Convert a melody from relative to absolute form by centering the mean of the stimuli on mean of the user's range.}
\usage{
rel_to_abs_mel_mean_centred(
  rel_melody,
  bottom_range,
  top_range,
  plot = FALSE,
  range = NULL,
  transpose = NULL
)
}
\arguments{
\item{transpose}{}
}
\description{
Convert a melody from relative to absolute form by centering the mean of the stimuli on mean of the user's range.
}
