% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{render_scores_table}
\alias{render_scores_table}
\title{Render scores in a shiny table}
\usage{
render_scores_table(res)
}
\arguments{
\item{res}{}
}
\value{

}
\description{
Render scores in a shiny table
}
