% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{rhythmic_melody_trials}
\alias{rhythmic_melody_trials}
\title{Present a block of rhythmic melody trials}
\usage{
rhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = NULL,
  feedback = FALSE,
  page_title = "Sing this melody plus rhythm"
)
}
\arguments{
\item{num_items}{}
}
\value{

}
\description{
Present a block of rhythmic melody trials
}
