% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{rhythmic_ntile_sampler_codeblock}
\alias{rhythmic_ntile_sampler_codeblock}
\title{Sample rhythmic melodies by N-tiles}
\usage{
rhythmic_ntile_sampler_codeblock(
  item_bank,
  num_items,
  id = "rhythmic_melody",
  phase = "test",
  n = 4
)
}
\arguments{
\item{n}{}
}
\description{
Sample rhythmic melodies by N-tiles
}
