% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_arrhythmic}
\alias{sample_arrhythmic}
\title{Sample arrhythmic}
\usage{
sample_arrhythmic(item_bank, num_items_arrhythmic, id = "arrhythmic_melody")
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Sample arrhythmic
}
