% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_from_user_range}
\alias{sample_from_user_range}
\title{Sample from the user's range (determined at test time)}
\usage{
sample_from_user_range(no_to_sample)
}
\arguments{
\item{no_to_sample}{}
}
\description{
Sample from the user's range (determined at test time)
}
