% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_in_ntiles_wrapper}
\alias{sample_in_ntiles_wrapper}
\title{Export sample in ntiles wrapper}
\usage{
sample_in_ntiles_wrapper(
  item_bank,
  num_items,
  n = 4,
  ntile_sampling_upper_bound = 4,
  col_name,
  phase = "test"
)
}
\arguments{
\item{phase}{}
}
\description{
Export sample in ntiles wrapper
}
