% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition.R
\name{sample_melody_in_key}
\alias{sample_melody_in_key}
\title{Sample melody in key}
\usage{
sample_melody_in_key(
  item_bank,
  inst,
  bottom_range,
  top_range,
  difficulty,
  length = NULL
)
}
\arguments{
\item{length}{}
}
\value{

}
\description{
Sample melody in key
}
