% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition.R
\name{sample_melody_in_key_v2}
\alias{sample_melody_in_key_v2}
\title{Sample melody in key (v2)}
\usage{
sample_melody_in_key_v2(
  item_bank,
  inst,
  bottom_range,
  top_range,
  difficulty,
  length = NULL
)
}
\arguments{
\item{length}{}
}
\description{
Sample melody in key (v2)
}
