% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_random}
\alias{sample_random}
\title{Randomly sample num_items per dataframe}
\usage{
sample_random(item_bank, num_items)
}
\arguments{
\item{num_items}{}
}
\description{
Randomly sample num_items per dataframe
}
