% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_review}
\alias{sample_review}
\title{Sample review}
\usage{
sample_review(
  num_review_items,
  id = "arrhythmic_melody_review",
  rhythmic = FALSE
)
}
\arguments{
\item{rhythmic}{}
}
\description{
Sample review
}
