% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_rhythmic}
\alias{sample_rhythmic}
\title{Sample rhythmic}
\usage{
sample_rhythmic(item_bank, num_items_rhythmic, id = "rhythmic_melody")
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Sample rhythmic
}
