% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_melodic_production}
\alias{score_melodic_production}
\title{Score using melodic production measures}
\usage{
score_melodic_production(
  user_melody_input = numeric(),
  user_duration_input = numeric(),
  user_onset_input = numeric(),
  stimuli = numeric(),
  stimuli_durations = numeric(),
  singing_measures = TRUE,
  pyin_pitch_track = tibble::tibble(),
  user_response_midi_note_off = numeric(),
  onsets_noteoff = numeric(),
  answer_meta_data = tibble::tibble()
)
}
\arguments{
\item{answer_meta_data}{}
}
\value{

}
\description{
Score using melodic production measures
}
