% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy_melody.R
\name{score_melody_interval_precision}
\alias{score_melody_interval_precision}
\title{Score melody interval precision}
\usage{
score_melody_interval_precision(all_participant_trials)
}
\arguments{
\item{all_participant_trials}{}
}
\description{
Score melody interval precision
}
