% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_rhythm_production.R
\name{score_rhythm_production}
\alias{score_rhythm_production}
\title{Score rhythm production}
\usage{
score_rhythm_production(stimuli_durations, user_durations, bpm = NULL)
}
\arguments{
\item{bpm}{}
}
\description{
Score rhythm production
}
