% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{segment_phrase}
\alias{segment_phrase}
\title{Segment a note track by adding phrasend and phrasbeg columns with boolean markers.}
\usage{
segment_phrase(note_track)
}
\arguments{
\item{note_track}{a data frame with an "onset" column}
}
\value{

}
\description{
Segment a note track by adding phrasend and phrasbeg columns with boolean markers.
}
