% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_midi_page.R
\name{select_midi_device_page}
\alias{select_midi_device_page}
\title{Create a page for selecting a MIDI device}
\usage{
select_midi_device_page(
  title = "Select MIDI device",
 
    message = "Your device should have been plugged in before you reached this page. It may take a moment to appear.",
  error_notification = "No midi device found.",
  button_text = psychTestR::i18n("Next")
)
}
\arguments{
\item{button_text}{}
}
\value{

}
\description{
Create a page for selecting a MIDI device
}
