% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{select_musical_instrument_page}
\alias{select_musical_instrument_page}
\title{Page to select what instrument the participant is using}
\usage{
select_musical_instrument_page()
}
\description{
Page to select what instrument the participant is using
}
