% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoken_words.R
\name{select_voice_page}
\alias{select_voice_page}
\title{A page for selecting a voice before a play words page}
\usage{
select_voice_page()
}
\description{
A page for selecting a voice before a play words page
}
