% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_condition_page}
\alias{set_condition_page}
\title{Allow the experimenter to set a condition at the beginning of the test}
\usage{
set_condition_page(block1, block2, condition1_name, condition2_name)
}
\arguments{
\item{condition2_name}{}
}
\value{

}
\description{
Allow the experimenter to set a condition at the beginning of the test
}
