% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_music.R
\name{set_default_range}
\alias{set_default_range}
\title{Get the default range for an instrument}
\usage{
set_default_range(instrument)
}
\arguments{
\item{instrument}{}
}
\description{
Get the default range for an instrument
}
