% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_init.R
\name{set_instrument}
\alias{set_instrument}
\title{Set instrument ID}
\usage{
set_instrument(
  instrument_id = NULL,
  as_code_block = TRUE,
  state = NULL,
  set_range = TRUE
)
}
\arguments{
\item{set_range}{}
}
\description{
Set instrument ID
}
