% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_utils.R
\name{set_instrument_range}
\alias{set_instrument_range}
\title{Set a dummy range (to avoid having to manually set in the test)}
\usage{
set_instrument_range(bottom_range = 48, top_range = 72)
}
\arguments{
\item{top_range}{}
}
\description{
Set a dummy range (to avoid having to manually set in the test)
}
