% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_music.R
\name{set_melodic_stimuli}
\alias{set_melodic_stimuli}
\title{Set melodic stimuli}
\usage{
set_melodic_stimuli(stimuli, durations, musicxml_file = "NA")
}
\arguments{
\item{musicxml_file}{}
}
\description{
Set melodic stimuli
}
