% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_musicassessr_state}
\alias{set_musicassessr_state}
\title{Set whether a musicassessr test is running in production or locally (test)}
\usage{
set_musicassessr_state(state = c(NULL, "production", "test"))
}
\arguments{
\item{state}{}
}
\value{

}
\description{
Set whether a musicassessr test is running in production or locally (test)
}
