% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_response_type}
\alias{set_response_type}
\title{Set response type for a test manually}
\usage{
set_response_type(type = c("Microphone", "MIDI"))
}
\arguments{
\item{type}{}
}
\description{
Set response type for a test manually
}
