% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_pages}
\alias{setup_pages}
\title{Setup pages for musicassessr test}
\usage{
setup_pages(
  input = c("microphone", "midi_keyboard", "midi_keyboard_and_microphone",
    "midi_keyboard_or_microphone"),
  headphones = TRUE,
  SNR_test = TRUE,
  min_SNR = 14,
  get_user_info = TRUE,
  demo = FALSE,
  get_instrument_range = FALSE,
  absolute_url = character(),
  select_instrument = FALSE,
  get_instrument_range_musical_notation = FALSE,
  adjust_range = FALSE,
  test_type = c("voice", "instrument"),
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  report_SNR = FALSE,
  concise_wording = FALSE
)
}
\arguments{
\item{input}{The form of musicassessr test.}

\item{headphones}{Deploy a headphone setup page.}

\item{SNR_test}{Deploy an SNR test.}

\item{min_SNR}{The minimum accepted SNR value.}

\item{get_user_info}{Get user info from their browser.}

\item{demo}{Deploy demo test.}

\item{get_instrument_range}{Get the user's pitch/instrument range at test time.}

\item{absolute_url}{If in production, what is the absolute URL.}

\item{select_instrument}{Deploy instrument selection page.}

\item{get_instrument_range_musical_notation}{When displaying instrument range, provide visual music notation.}

\item{adjust_range}{Adjust range.}

\item{test_type}{Voice vs. instrument test.}

\item{microphone_test}{Deploy microphone test page.}

\item{allow_repeat_SNR_tests}{Allow repeated SNR tests, if FALSE, then participant only gets one go and the test will fail if the SNR test fails.}

\item{report_SNR}{Should the SNR be reported to the user.}

\item{concise_wording}{Whether the wording should be concise or not.}
}
\value{

}
\description{
Setup pages for musicassessr test
}
