% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21.R
\name{sing_happy_birthday_page}
\alias{sing_happy_birthday_page}
\title{A page to identify a user's singing range by asking them to sing Happy Birthday}
\usage{
sing_happy_birthday_page(feedback = FALSE)
}
\arguments{
\item{feedback}{}
}
\value{

}
\description{
A page to identify a user's singing range by asking them to sing Happy Birthday
}
