% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{sing_rhythmic_melody_trials}
\alias{sing_rhythmic_melody_trials}
\title{Sing rhythmic melody trial block}
\usage{
sing_rhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production_additional_measures,
  sound = "piano",
  page_text = psychTestR::i18n("sing_melody_trial"),
  page_title = "Sing This Melody Plus Rhythm",
  page_type = "record_audio_page",
 
    instruction_text = "Now you will hear melodies with rhythms. Please try and sing the melodies with the correct rhythm.",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_record_button = FALSE,
  show_progress = TRUE,
  module_name = "sing_rhythmic_melodies"
)
}
\arguments{
\item{module_name}{}
}
\value{

}
\description{
Sing rhythmic melody trial block
}
