% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songbird.R
\name{songbird_welcome_page_fun}
\alias{songbird_welcome_page_fun}
\title{Songbird welcome}
\usage{
songbird_welcome_page_fun(success, username)
}
\arguments{
\item{username}{}
}
\description{
Songbird welcome
}
