% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_string_df}
\alias{to_string_df}
\title{Useful (but unsophisticated) util to collapse a df into a pretty string df}
\usage{
to_string_df(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Useful (but unsophisticated) util to collapse a df into a pretty string df
}
