% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{translate_from_dict}
\alias{translate_from_dict}
\title{Translate an item from musicassessr dictionary}
\usage{
translate_from_dict(non_english_translation, language)
}
\arguments{
\item{language}{}
}
\value{

}
\description{
Translate an item from musicassessr dictionary
}
