% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition.R
\name{transpose_melody_to_easy_or_hard_key}
\alias{transpose_melody_to_easy_or_hard_key}
\title{Transpose a melody to a common/easy or uncommon/hard key}
\usage{
transpose_melody_to_easy_or_hard_key(
  abs_melody,
  difficulty,
  inst,
  bottom_range,
  top_range
)
}
\arguments{
\item{top_range}{}
}
\description{
Transpose a melody to a common/easy or uncommon/hard key
}
