% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition.R
\name{transpose_melody_to_key}
\alias{transpose_melody_to_key}
\title{Transpose a melody to a given key}
\usage{
transpose_melody_to_key(abs_melody, key, bottom_range, top_range)
}
\arguments{
\item{top_range}{}
}
\description{
Transpose a melody to a given key
}
