% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_melody_loop.R
\name{update_play_melody_loop_and_save}
\alias{update_play_melody_loop_and_save}
\title{Update play melody loop and save}
\usage{
update_play_melody_loop_and_save(max_goes)
}
\arguments{
\item{max_goes}{}
}
\description{
Update play melody loop and save
}
