% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21.R
\name{UPEI_2021_battery}
\alias{UPEI_2021_battery}
\title{UPEI Battery}
\usage{
UPEI_2021_battery(
  state = "production",
  aws_credentials = list(api_url =
    "https://255uxe6ajl.execute-api.us-east-1.amazonaws.com/api", bucket_name =
    "shinny-app-source-41630", bucket_region = "us-east-1", identity_pool_id =
    "us-east-1:feecdf7e-cdf6-416f-94d0-a6de428c8c6b", destination_bucket =
    "shinny-app-destination-41630")
)
}
\arguments{
\item{aws_credentials}{}
}
\value{

}
\description{
UPEI Battery
}
