% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_deprecated.R
\name{user_info_check}
\alias{user_info_check}
\title{user_info_check}
\arguments{
\item{input}{Input from the user.}

\item{state}{Current state of the user.}

\item{...}{Additional arguments.}
}
\value{
A list containing the user information if the browser is capable, otherwise an error is displayed.
}
\description{
Check the browser information and save it including participant ID.
}
