% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_video.R
\name{video_melodic_production_page}
\alias{video_melodic_production_page}
\title{Video melodic production page}
\usage{
video_melodic_production_page(video_file, melody_no, total_no_melodies)
}
\arguments{
\item{total_no_melodies}{}
}
\description{
Video melodic production page
}
