% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{wait_for_api_page}
\alias{wait_for_api_page}
\title{Wait for API page}
\usage{
wait_for_api_page(
  poll_frequency_seconds = 1L,
  check_function = check_session_id_ready
)
}
\arguments{
\item{check_function}{This should return TRUE if not ready (still need to wait for API) or FALSE if ready (api complete).}
}
\description{
Wait for API page
}
