% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{wjd_audio_melody_trials}
\alias{wjd_audio_melody_trials}
\title{A trial block which plays back real audio from the Weimar Jazz Database}
\usage{
wjd_audio_melody_trials(
  item_bank,
  num_items,
  num_examples = FALSE,
  feedback = FALSE,
  get_answer = musicassessr::get_answer_pyin,
  sound = "piano",
 
    page_text = "Press play to hear the melody, then play it back as best as you can when it finishes.",
  page_title = "Play the Melody And Rhythm from Audio",
 
    instruction_text = "Now you will hear some melodies as audio. Please try and play back the melodies and rhythms as best as you can."
)
}
\arguments{
\item{instruction_text}{}
}
\value{

}
\description{
A trial block which plays back real audio from the Weimar Jazz Database
}
