% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradigms.R
\name{wrap_js_fun_body}
\alias{wrap_js_fun_body}
\title{Helper to wrap some JS code in a function}
\usage{
wrap_js_fun_body(js_code)
}
\arguments{
\item{js_code}{}
}
\description{
Helper to wrap some JS code in a function
}
