% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{arrhythmic_melody_trials}
\alias{arrhythmic_melody_trials}
\title{Arrhythmic Melody Trials Block}
\usage{
arrhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = FALSE,
  feedback = FALSE,
  get_answer = musicassessr::get_answer_pyin,
  sound = "piano",
  page_text = psychTestR::i18n("sing_melody_trial"),
  page_title = "Sing the Melody",
 
    instruction_text = "Now you will hear some melodies. Please try and sing the melodies."
)
}
\arguments{
\item{instruction_text}{}
}
\value{

}
\description{
Arrhythmic Melody Trials Block
}
