% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{arrhythmic_melody_trials}
\alias{arrhythmic_melody_trials}
\title{Arrhythmic Melody Trials Block}
\usage{
arrhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production_additional_measures,
  sound = "piano",
 
    page_text = "Click below to hear the melody. Play back the melody. Click Stop when finished.",
  page_title = "Play the Melody",
  page_type = "record_audio_page",
 
    instruction_text = "Now you will hear some melodies. Please try and play the melodies.",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_record_button = FALSE,
  show_progress = TRUE,
  module_name = "arrhythmic_melodies",
  start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  show_sheet_music = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  sampler_function = sample_arrhythmic,
  presampled = FALSE
)
}
\arguments{
\item{presampled}{}
}
\value{

}
\description{
Arrhythmic Melody Trials Block
}
