% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{arrhythmic_melody_trials}
\alias{arrhythmic_melody_trials}
\title{Arrhythmic melody trials block}
\usage{
arrhythmic_melody_trials(
  var_name = "arrhythmic_melody",
  module_name = "arrhythmic_melodies",
  page_text = "Press play to hear the melody. Play back the melody.",
  page_title = "Play this melody.",
  instruction_text =
    "Now you will hear some melodies. Please try and play back the melodies.",
  sampler_function = sample_arrhythmic,
  item_bank,
  num_items = integer(),
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production_additional_measures,
  sound = "piano",
  page_type = "record_audio_page",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_progress = TRUE,
  start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  show_sheet_music = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  presampled = FALSE,
  get_similarity_to_previous_melody = FALSE,
  volume_meter = FALSE,
  volume_meter_type = "default",
  melody_paradigm = c("standard", "sing_melody_first",
    "learn_phase_visual_display_modality"),
  singing_trials = TRUE,
  review = FALSE,
  phase = c("test", "learn", "review")
)
}
\arguments{
\item{phase}{}
}
\description{
Arrhythmic melody trials block
}
