% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{arrhythmic_melody_trials}
\alias{arrhythmic_melody_trials}
\title{Arrhythmic Melody Trials Block}
\usage{
arrhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = NULL,
  feedback = FALSE
)
}
\arguments{
\item{feedback}{}
}
\value{

}
\description{
Arrhythmic Melody Trials Block
}
