% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{audio_melodic_production_trials}
\alias{audio_melodic_production_trials}
\title{Present a trial block of melodies from audio files}
\usage{
audio_melodic_production_trials(
  audio_directory,
  no_to_sample = NULL,
  module_prefix = "audio_trial",
  page_title = "Sing back the melody",
 
    page_text = "Click below to hear the melody. Sing back the melody. Click Stop when finished.",
  grab_meta_data,
  meta_data_df,
  meta_data_lookup_column,
  get_answer = musicassessr::get_answer_pyin_melodic_production,
  feedback = FALSE
)
}
\arguments{
\item{feedback}{}
}
\value{

}
\description{
Present a trial block of melodies from audio files
}
