% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{audio_melodic_production_trials}
\alias{audio_melodic_production_trials}
\title{Present a trial block of melodies from audio files}
\usage{
audio_melodic_production_trials(
  audio_directory,
  no_to_sample = NULL,
  module_prefix = "audio_trial",
  page_title = psychTestR::i18n("Sing_the_Melody"),
  page_text = psychTestR::i18n("sing_melody_page_text"),
  grab_meta_data,
  meta_data_df,
  meta_data_lookup_column,
  get_answer = get_answer_pyin_melodic_production_additional_measures,
  feedback = FALSE
)
}
\arguments{
\item{feedback}{}
}
\description{
Present a trial block of melodies from audio files
}
