% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_midi_or_audio_ui.R
\name{create_db_vars_template}
\alias{create_db_vars_template}
\title{Create a template for db_vars}
\usage{
create_db_vars_template(
  init_with_time_started = TRUE,
  trial_time_started = Sys.time(),
  trial_paradigm = "call_and_response",
  onset = FALSE,
  pyin_type = "notes",
  attempt = 1L,
  item_id = NULL,
  page_label = NULL,
  display_modality = "auditory",
  phase = "test",
  feedback = FALSE,
  additional = list(),
  test_id = 1L,
  session_id = NULL,
  instrument = NULL,
  melody_block_paradigm = "NA",
  module = "NA",
  rhythmic = TRUE,
  user_id = NULL,
  feedback_type = "opti3"
)
}
\arguments{
\item{user_id}{}
}
\description{
Create a template for db_vars
}
