% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{empty_page}
\alias{empty_page}
\title{An empty page (with trigger button hidden and possible to show on demand.)}
\usage{
empty_page(
  body = "",
  admin_ui = NULL,
  button_text = "Next",
  on_complete = NULL,
  page_title = ""
)
}
\arguments{
\item{page_title}{}
}
\value{

}
\description{
An empty page (with trigger button hidden and possible to show on demand.)
}
