% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_long_note_screening.R
\name{end_of_long_note_trial_screening}
\alias{end_of_long_note_trial_screening}
\title{End of long note trial screening code block}
\usage{
end_of_long_note_trial_screening(failure_page = "http://www.google.com")
}
\value{

}
\description{
End of long note trial screening code block
}
