% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{final_page_or_continue_to_new_test}
\alias{final_page_or_continue_to_new_test}
\title{Should the end of the test present a final page or not?}
\usage{
final_page_or_continue_to_new_test(
  final = TRUE,
  task_name,
  img = NULL,
  present_continue_to_new_test_page = TRUE,
  redirect_url = NULL
)
}
\arguments{
\item{redirect_url}{}
}
\description{
Should the end of the test present a final page or not?
}
