% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{final_results}
\alias{final_results}
\title{Present final results}
\usage{
final_results(
  test_name,
  url,
  num_items_long_tone,
  num_items_arrhythmic,
  num_items_rhythmic,
  socials = FALSE
)
}
\arguments{
\item{socials}{}
}
\value{

}
\description{
Present final results
}
