% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{find_this_note_trials}
\alias{find_this_note_trials}
\title{Present "Find This Note" trials}
\usage{
find_this_note_trials(
  num_items,
  num_examples = NULL,
  feedback = FALSE,
  page_title = "Find This Note",
  get_answer = get_answer_pyin,
  page_type = "record_audio_page",
 
    page_text = "Press play to hear the note. Try and play it on your instrument when you can."
)
}
\arguments{
\item{page_type}{}
}
\value{

}
\description{
Present "Find This Note" trials
}
