% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_onset_detection}
\alias{get_answer_onset_detection}
\title{Get onset information from an audio file}
\usage{
get_answer_onset_detection(input, state, ...)
}
\description{
Get onset information from an audio file
}
