% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin}
\alias{get_answer_pyin}
\title{Use pyin on a file}
\usage{
get_answer_pyin(input, type = c("both", "note", "pitch_track"), state, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Use pyin on a file
}
