% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer_pyin}
\alias{get_answer_pyin}
\title{Get answer by running pyin on a recorded audio file}
\usage{
get_answer_pyin(
  input,
  type = c("both", "note", "pitch_track"),
  state,
  melconv = FALSE,
  singing_measures = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Get answer by running pyin on a recorded audio file
}
