% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{get_implicit_harmonies}
\alias{get_implicit_harmonies}
\title{get harmonies via the Krumhansl-Schmuckler algorithm}
\usage{
get_implicit_harmonies(
  pitch_vec,
  segmentation = NULL,
  only_winner = TRUE,
  segmentation_type = c("phrase_boundary_marker", "segment_id")
)
}
\arguments{
\item{segmentation_type}{}
}
\value{

}
\description{
get harmonies via the Krumhansl-Schmuckler algorithm
}
