% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_builders.R
\name{get_instrument_range_pages}
\alias{get_instrument_range_pages}
\title{Get Instrument Range Pages}
\usage{
get_instrument_range_pages(type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Get Instrument Range Pages
}
