% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_range.R
\name{get_instrument_range_pages}
\alias{get_instrument_range_pages}
\title{Get Instrument Range Pages}
\usage{
get_instrument_range_pages(
  type,
  get_range,
  show_musical_notation = FALSE,
  adjust_range = FALSE,
  test_type = c("voice", "instrument"),
  concise_wording = FALSE,
  default_range = list(bottom_range = 48, top_range = 72)
)
}
\arguments{
\item{default_range}{}
}
\description{
Get Instrument Range Pages
}
