% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_info.R
\name{get_user_info_page}
\alias{get_user_info_page}
\title{Check a user has the requirements to run our test and record their browser information}
\usage{
get_user_info_page(chrome_only = TRUE)
}
\arguments{
\item{chrome_only}{}
}
\description{
Check a user has the requirements to run our test and record their browser information
}
