% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{include_musicassessr_js}
\alias{include_musicassessr_js}
\title{Include musicassessr scripts in a webpage}
\usage{
include_musicassessr_js(
  app_name = "",
  visual_notation = FALSE,
  record_audio = TRUE,
  midi_input = FALSE
)
}
\arguments{
\item{midi_input}{}
}
\description{
Include musicassessr scripts in a webpage
}
