% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{interval_perception_trials}
\alias{interval_perception_trials}
\title{A block which test a participant's perception of intervals}
\usage{
interval_perception_trials(
  num_items = 26L,
  sound = "piano",
  page_title = "What's that interval?",
  instruction_text =
    shiny::tags$div(shiny::tags$p("In the next set of trials, you will hear a musical interval."),
    shiny::tags$p("You must try and say what the interval is."),
    shiny::tags$p("There are no practice rounds, you will begin immediately."))
)
}
\arguments{
\item{instruction_text}{}
}
\description{
A block which test a participant's perception of intervals
}
