% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_singing_accuracy.R
\name{long_note_pitch_metrics}
\alias{long_note_pitch_metrics}
\title{Get pitch metrics for long tone trials}
\usage{
long_note_pitch_metrics(target_pitch, freq, state)
}
\arguments{
\item{freq}{}
}
\value{

}
\description{
Get pitch metrics for long tone trials
}
