% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{long_tone_trials}
\alias{long_tone_trials}
\title{Present a block of long tone trials}
\usage{
long_tone_trials(
  num_items,
  num_examples = 0L,
  feedback = FALSE,
  show_instructions = TRUE,
  page_text = psychTestR::i18n("long_tone_text"),
  page_title = psychTestR::i18n("long_tone_title"),
  get_answer = get_answer_pyin_long_note,
  page_type = "record_audio_page",
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  instruction_text = shiny::div(shiny::tags$h2(page_title),
    shiny::tags$p(psychTestR::i18n("long_tone_instruction")),
    shiny::tags$p(psychTestR::i18n("long_tone_instruction_2")),
    shiny::tags$p(psychTestR::i18n("long_tone_instruction_3"))),
  module_name = "long_tone_trials",
  show_progress = TRUE
)
}
\arguments{
\item{show_progress}{}
}
\value{

}
\description{
Present a block of long tone trials
}
