% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{long_tone_trials}
\alias{long_tone_trials}
\title{Present a block of long tone trials}
\usage{
long_tone_trials(
  num_items,
  num_examples = NULL,
  feedback = FALSE,
  page_title = "Sing Along With This Note",
  get_answer = musicassessr::get_answer_save_aws_key
)
}
\arguments{
\item{get_answer}{}
}
\value{

}
\description{
Present a block of long tone trials
}
