% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_musicassessr_test.R
\name{make_musicassessr_test}
\alias{make_musicassessr_test}
\title{Make a musicassessr test}
\usage{
make_musicassessr_test(
  musicassessr_state = "production",
  elts_before_setup_pages = function() {    
    psychTestR::join(psychTestR::code_block(function(state, ...) {     })) },
  elts,
  setup_pages,
  setup_pages_options = musicassessr::setup_pages_options,
  title,
  admin_password,
  languages = "en",
  additional_dict = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Make a musicassessr test
}
