% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_musicassessr_test.R
\name{make_musicassessr_test}
\alias{make_musicassessr_test}
\title{Make a musicassessr test}
\usage{
make_musicassessr_test(
  title,
  admin_password,
  elts,
  elts_before_setup_pages = function() {
     empty_code_block()
 },
  languages = c("en", "it", "de", "lv"),
  opt = musicassessr_opt(),
  final_page =
    wrap_musicassessr_timeline(psychTestR::final_page(psychTestR::i18n("thank_you_for_completing"))),
  welcome_page = psychTestR::one_button_page("Welcome."),
  dict = musicassessr::musicassessr_dict,
  ...
)
}
\arguments{
\item{title}{Title of the test.}

\item{admin_password}{Password for the test.}

\item{elts}{Timeline to go after setup pages.}

\item{elts_before_setup_pages}{Timeline to go before setup pages.}

\item{languages}{Languages for the test.}

\item{opt}{Musicassessr options.}

\item{final_page}{The UI of the final page.}

\item{welcome_page}{Required because you need a page before musicassessr_init to instantiate a p_id.}

\item{...}{}
}
\description{
Make a musicassessr test
}
