% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{melody_trials}
\alias{melody_trials}
\title{Melody trials constructor}
\usage{
melody_trials(
  var_name,
  module_name,
  page_text,
  page_title,
  instruction_text,
  sampler_function,
  item_bank,
  num_items = integer(),
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production_additional_measures,
  sound = "piano",
  page_type = "record_audio_page",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_progress = TRUE,
  sheet_music_start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  presampled = FALSE,
  arrhythmic = FALSE,
  get_similarity_to_previous_melody = FALSE,
  volume_meter = FALSE,
  volume_meter_type = "default",
  melody_block_paradigm = c("standard", "sing_melody_first",
    "learn_phase_visual_display_modality"),
  singing_trials = TRUE,
  review = FALSE,
  phase = c("test", "learn", "review")
)
}
\arguments{
\item{phase}{}
}
\description{
Melody trials constructor
}
