% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microphone_calibration_page.R
\name{microphone_calibration_page}
\alias{microphone_calibration_page}
\title{Microphone Calibration Page}
\usage{
microphone_calibration_page(
  button_text = "Start test",
  concise_wording = FALSE,
  musical_instrument = FALSE
)
}
\arguments{
\item{musical_instrument}{}
}
\value{

}
\description{
Microphone Calibration Page
}
