% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microphone_calibration_page.R
\name{microphone_calibration_page}
\alias{microphone_calibration_page}
\title{Microphone Calibration Page}
\usage{
microphone_calibration_page(button_text = "Start test")
}
\arguments{
\item{deploy_crepe_stats}{}
}
\value{

}
\description{
Microphone Calibration Page
}
