% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_range.R
\name{multi_page_play_melody_loop}
\alias{multi_page_play_melody_loop}
\title{A page builder for creating a specified number of play_melody_loops}
\usage{
multi_page_play_melody_loop(
  n_items,
  var_name = "melody",
  stimuli_type = "midi_notes",
  page_type = "record_audio_page",
  max_goes = 3,
  page_title = psychTestR::i18n("copy_melody_title"),
 
    page_text = "Press play to hear the melody, then play it back as best as you can when it finishes.",
  get_answer = get_answer_save_aws_key,
  rel_to_abs_mel_function = musicassessr::rel_to_abs_mel_mean_centred,
  start_from_trial_no = 1,
  clip_stimuli_length = FALSE,
  arrhythmic = FALSE,
  example = FALSE,
  feedback = FALSE
)
}
\arguments{
\item{feedback}{}
}
\value{

}
\description{
A page builder for creating a specified number of play_melody_loops
}
