% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_melody_loop.R
\name{multi_page_play_melody_loop}
\alias{multi_page_play_melody_loop}
\title{A page builder for creating a specified number of play_melody_loops}
\usage{
multi_page_play_melody_loop(
  presampled_items = NULL,
  n_items,
  var_name = "melody",
  stimuli_type = "midi_notes",
  page_type = "record_audio_page",
  max_goes = 3L,
  page_title = psychTestR::i18n("copy_melody_title"),
 
    page_text = "Press play to hear the melody, then play it back as best as you can when it finishes.",
  get_answer = get_answer_pyin_melodic_production,
  rel_to_abs_mel_function = NULL,
  start_from_trial_no = 1L,
  clip_stimuli_length = FALSE,
  arrhythmic = FALSE,
  example = FALSE,
  feedback = FALSE,
  sound = "piano",
  get_trial_characteristics_function = NULL,
  max_goes_forced = FALSE,
  give_first_melody_note = FALSE
)
}
\arguments{
\item{give_first_melody_note}{}
}
\value{

}
\description{
A page builder for creating a specified number of play_melody_loops
}
