% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_long_tone_record_audio.R
\name{multi_play_long_tone_record_audio_pages}
\alias{multi_play_long_tone_record_audio_pages}
\title{Build multiple play_long_tone_record_audio_pages based on a user's range}
\usage{
multi_play_long_tone_record_audio_pages(
  no_items,
  page_type = "record_audio_page",
  page_text = psychTestR::i18n("long_tone_text"),
  example = FALSE,
  page_title = psychTestR::i18n("long_tone_title"),
  feedback = FALSE,
  get_answer = get_answer_pyin_long_note,
  show_progress = TRUE,
  trial_paradigm = c("simultaneous_recall", "call_and_response"),
  call_and_response_end = c("manual", "auto"),
  long_tone_length = 5,
  singing_trial = FALSE,
  mute_midi_playback = FALSE,
  volume_meter = FALSE,
  volume_meter_type = "default"
)
}
\arguments{
\item{volume_meter_type}{}
}
\description{
Build multiple play_long_tone_record_audio_pages based on a user's range
}
