% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_range.R,
%   R/play_long_tone_record_audio.R
\name{multi_play_long_tone_record_audio_pages}
\alias{multi_play_long_tone_record_audio_pages}
\title{Build multiple play_long_tone_record_audio_pages based on a user's range}
\usage{
multi_play_long_tone_record_audio_pages(
  no_items,
  page_type = "record_audio_page",
  example = FALSE,
  feedback = FALSE,
  get_answer = get_answer_pyin_long_note
)

multi_play_long_tone_record_audio_pages(
  no_items,
  page_type = "record_audio_page",
  example = FALSE,
  feedback = FALSE,
  get_answer = get_answer_pyin_long_note
)
}
\arguments{
\item{feedback}{}
}
\value{



}
\description{
Build multiple play_long_tone_record_audio_pages based on a user's range

Build multiple play_long_tone_record_audio_pages based on a user's range
}
