% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_init.R
\name{musicassessr_init}
\alias{musicassessr_init}
\title{Initiate a musicassessr test}
\usage{
musicassessr_init(
  use_musicassessr_db = FALSE,
  app_name = "",
  experiment_id = NULL,
  experiment_condition_id = NULL,
  user_id = NULL
)
}
\arguments{
\item{user_id}{}
}
\description{
Initiate a musicassessr test
}
