% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_init.R
\name{musicassessr_init}
\alias{musicassessr_init}
\title{Initiate a musicassessr test}
\usage{
musicassessr_init(
  test_username,
  test,
  store_results_in_db = FALSE,
  local_app_file_dir,
  sonic_annotator_local_location
)
}
\arguments{
\item{sonic_annotator_local_location}{}
}
\value{

}
\description{
Initiate a musicassessr test
}
