% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_init.R
\name{musicassessr_init}
\alias{musicassessr_init}
\title{Initiate a musicassessr test}
\usage{
musicassessr_init(
  test_username = NA,
  test = NA,
  store_results_in_db = FALSE,
  local_app_file_dir = "/Users/sebsilas/aws-musicassessr-local-file-upload/files/",
  sonic_annotator_local_location = "/Users/sebsilas/sonic-annotator",
  copy_audio_to_location = NULL
)
}
\arguments{
\item{copy_audio_to_location}{}
}
\value{

}
\description{
Initiate a musicassessr test
}
