% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_init.R
\name{musicassessr_init}
\alias{musicassessr_init}
\title{Initiate a musicassessr test}
\usage{
musicassessr_init(
  test_username = NA,
  test = NA,
  store_results_in_db = FALSE,
  app_name
)
}
\arguments{
\item{app_name}{}
}
\value{

}
\description{
Initiate a musicassessr test
}
