% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{musicassessr_js}
\alias{musicassessr_js}
\title{musicassessr scripts}
\usage{
musicassessr_js(
  app_name = "",
  asynchronous_api_mode = FALSE,
  visual_notation = FALSE,
  midi_file_playback = FALSE,
  record_audio = TRUE,
  midi_input = FALSE
)
}
\arguments{
\item{midi_input}{}
}
\description{
musicassessr scripts
}
