% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{musicassessr_js}
\alias{musicassessr_js}
\title{musicassessr scripts}
\usage{
musicassessr_js(
  musicassessr_aws = FALSE,
  visual_notation = FALSE,
  midi_file_playback = FALSE,
  record_audio = TRUE,
  app_name = character()
)
}
\arguments{
\item{app_name}{}
}
\value{

}
\description{
musicassessr scripts
}
