% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicassessr_js_scripts.R
\name{musicassessr_js_scripts}
\alias{musicassessr_js_scripts}
\title{musicassessr_js_scripts}
\usage{
musicassessr_js_scripts(
  destination_bucket,
  api_url,
  bucket_name,
  bucket_region,
  identity_pool_id,
  musicassessr_state = "production"
)
}
\arguments{
\item{identity_pool_id}{}
}
\value{

}
\description{
musicassessr_js_scripts
}
