% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_musicassessr_test.R
\name{musicassessr_opt}
\alias{musicassessr_opt}
\title{Specify options for a musicassessr test.}
\usage{
musicassessr_opt(
  test_username = NA,
  test = NA,
  store_results_in_db = FALSE,
  local_app_file_dir = "/Users/sebsilas/aws-musicassessr-local-file-upload/files/",
  sonic_annotator_local_location = "/Users/sebsilas/sonic-annotator",
  copy_audio_to_location = NULL
)
}
\arguments{
\item{copy_audio_to_location}{}
}
\value{

}
\description{
Specify options for a musicassessr test.
}
