% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_musicassessr_test.R
\name{musicassessr_opt}
\alias{musicassessr_opt}
\title{Specify options for a musicassessr test.}
\usage{
musicassessr_opt(
  test_username = NA,
  test = NA,
  store_results_in_db = FALSE,
  app_name = ""
)
}
\arguments{
\item{app_name}{}
}
\value{

}
\description{
Specify options for a musicassessr test.
}
