% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_simile.R
\name{opti3}
\alias{opti3}
\title{Score using the opti3 measure of similarity}
\usage{
opti3(
  pitch_vec1,
  onset_vec1,
  pitch_vec2,
  onset_vec2,
  N = 3,
  use_bootstrap = TRUE,
  return_components = FALSE,
  segmentation1 = NULL,
  segmentation2 = NULL
)
}
\arguments{
\item{segmentation2}{}
}
\description{
Score using the opti3 measure of similarity
}
