% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradigms.R
\name{paradigm}
\alias{paradigm}
\title{Choose a melody paradigm}
\usage{
paradigm(
  paradigm_type = c("call_and_response", "simultaneous_recall"),
  page_type = c("record_audio_page", "record_midi_page"),
  call_and_response_end = c("manual", "auto"),
  stop_recording_after_x_seconds = NULL,
  instantiate_midi = FALSE,
  midi_device = NULL,
  mute_midi_playback = FALSE
)
}
\arguments{
\item{mute_midi_playback}{}
}
\description{
Choose a melody paradigm
}
