% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_long_tone_record_audio.R
\name{play_long_tone_record_audio_page}
\alias{play_long_tone_record_audio_page}
\title{Create a page which produces a given tone}
\usage{
play_long_tone_record_audio_page(
  note = NULL,
  long_note_no = "x",
  note_length = 5,
  page_title = psychTestR::i18n("long_tone_heading"),
  page_text = "Sing along with the tone for 5 seconds.",
  play_button_text = "Play",
  page_type = "record_audio_page",
  show_aws_controls = FALSE,
  show_record_button = FALSE,
  auto_next_page = TRUE,
  example = FALSE,
  get_answer = get_answer_pyin_long_note,
  page_label = "long_tone_"
)
}
\arguments{
\item{get_answer}{}
}
\value{

}
\description{
Create a page which produces a given tone
}
