% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_melody_loop.R
\name{play_melody_loop}
\alias{play_melody_loop}
\title{Create a psychTestR test loop for having several attempts at playing back a melody.}
\usage{
play_melody_loop(
  melody = NULL,
  melody_no = "x",
  var_name = "melody",
  stimuli_type = "midi_notes",
  max_goes = 3,
  page_type = "record_audio_page",
  page_title = "Copy The Melody",
 
    page_text = "Press play to hear the melody, then play it back as best as you can when it finishes.",
  answer_meta_data = " ",
  get_answer = get_answer_pyin,
  rel_to_abs_mel_function = musicassessr:rel_to_abs_mel_mean_centred,
  clip_stimuli_length = FALSE,
  start_note = 1,
  end_note = "end",
  dur_list = "null",
  arrhythmic = FALSE,
  note_length = 0.5,
  play_button_text = psychTestR::i18n("Play"),
  example = FALSE,
  sound = "piano"
)
}
\arguments{
\item{example}{}
}
\value{

}
\description{
Create a psychTestR test loop for having several attempts at playing back a melody.
}
