% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_melody_loop.R
\name{play_melody_loop}
\alias{play_melody_loop}
\title{Create a psychTestR test loop for having several attempts at playing back a melody.}
\usage{
play_melody_loop(
  item_bank = NULL,
  melody = NULL,
  melody_no = 0,
  var_name = "melody",
  stimuli_type = "midi_notes",
  max_goes = 4L,
  max_goes_forced = FALSE,
  page_type = "record_audio_page",
  page_title = "Copy The Melody",
  page_text =
    "Press play to hear the melody, then play it back as best as you can when it finishes.",
  answer_meta_data = data.frame(),
  get_answer = get_answer_pyin_melodic_production,
  rel_to_abs_mel_function = NULL,
  clip_stimuli_length = FALSE,
  start_note = 1L,
  end_note = "end",
  durations = "null",
  arrhythmic = FALSE,
  note_length = 0.5,
  play_button_text = psychTestR::i18n("Play"),
  example = FALSE,
  sound = "piano",
  reactive_stimuli = NULL,
  get_trial_characteristics_function = NULL,
  display_modality = "auditory",
  total_no_melodies = 0,
  show_progress = TRUE,
  sheet_music_start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  psychTestRCAT = FALSE,
  melody_row = NULL,
  get_similarity_to_previous_melody = FALSE,
  volume_meter = FALSE,
  volume_meter_type = "default",
  singing_trial = TRUE,
  phase = c("test", "learn", "review", "example"),
  melody_block_paradigm = c("standard", "sing_melody_first",
    "learn_phase_visual_display_modality"),
  melody_trial_paradigm = c("call_and_response", "simultaneous_recall"),
  first_note_message = psychTestR::i18n("first_note_is"),
  transposed_message = psychTestR::i18n("transposed"),
  play_first_note_button_text = psychTestR::i18n("play_first_note"),
  skip_sampling_and_take_from_last_melody = FALSE,
  reactive_melody_no = FALSE,
  learn_test_paradigm = FALSE
)
}
\arguments{
\item{singing_trial}{Boolean: if TRUE, the trial is defined as singing-based, otherwise instrument-based.}

\item{phase}{Can be one of 'learn', 'test', 'review' or 'example'}

\item{learn_test_paradigm}{}
}
\description{
Create a psychTestR test loop for having several attempts at playing back a melody.
}
