% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_normal_dist_plus_score}
\alias{plot_normal_dist_plus_score}
\title{Plot a normal distribution with a certain score highlighted}
\usage{
plot_normal_dist_plus_score(
  data = NULL,
  mean = NULL,
  sd = NULL,
  highlighted_score
)
}
\arguments{
\item{highlighted_score}{}
}
\value{

}
\description{
Plot a normal distribution with a certain score highlighted
}
