% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli.R
\name{present_stimuli}
\alias{present_stimuli}
\title{Present stimuli}
\usage{
present_stimuli(
  stimuli,
  stimuli_type,
  display_modality,
  page_type = character(),
  page_text = " ",
  page_title = " ",
  slide_length = numeric(),
  answer_meta_data = character(),
  get_answer = function() { },
  save_answer = TRUE,
  stimuli_reactive = FALSE,
  midi_device = " ",
  show_aws_controls = FALSE,
  page_label = "x",
  button_text = "Next",
  play_button_text = "Play",
  note_length = 0.5,
  sound = "piano",
  asChord = FALSE,
  ascending = TRUE,
  start_note = 0L,
  end_note = "end",
  durations = numeric(),
  show_record_button = FALSE,
  auto_next_page = FALSE,
  choices = character(),
  user_rating = FALSE,
  page_text_first = TRUE,
  happy_with_response = FALSE,
  attempts_left = integer(),
  visual_music_notation_id = "sheet_music",
  play_button_id = "playButton",
  button_area_id = "button_area",
  hideOnPlay = FALSE,
  record_immediately = FALSE,
  max_goes_forced = FALSE,
  transpose_first_melody_note = 0,
  clef = "auto",
  volume = 1,
  audio_playback_as_single_play_button = FALSE,
  max_goes = 1,
  melody_no = 0,
  total_no_melodies = 0,
  show_progress = FALSE,
  start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  show_sheet_music = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  slider_value = 5,
  slider_min = 1,
  slider_max = 10,
  octave = 4,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Present stimuli
}
