% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli.R
\name{present_stimuli}
\alias{present_stimuli}
\title{Present Stimuli}
\usage{
present_stimuli(
  stimuli,
  stimuli_type,
  display_modality,
  page_type = character(),
  page_text = " ",
  page_title = " ",
  slide_length = numeric(),
  answer_meta_data = character(),
  get_answer = get_answer_null,
  save_answer = TRUE,
  stimuli_reactive = FALSE,
  midi_device = " ",
  show_aws_controls = FALSE,
  page_label = "x",
  button_text = "Next",
  play_button_text = "Play",
  note_length = 0.5,
  sound = "piano",
  asChord = FALSE,
  ascending = TRUE,
  start_note = 0L,
  end_note = "end",
  durations = numeric(),
  show_record_button = FALSE,
  auto_next_page = FALSE,
  choices = character(),
  user_rating = FALSE,
  page_text_first = TRUE,
  happy_with_response = FALSE,
  attempts_left = integer(),
  visual_music_notation_id = "sheet_music",
  play_button_id = "playButton",
  button_area_id = "button_area",
  hideOnPlay = FALSE,
  record_immediately = FALSE,
  max_goes_forced = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Present Stimuli
}
