% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli.R
\name{present_stimuli}
\alias{present_stimuli}
\title{Present Stimuli}
\usage{
present_stimuli(
  stimuli,
  stimuli_type,
  display_modality,
  page_type = NULL,
  page_text = " ",
  page_title = " ",
  slide_length,
  special_page_underlying_page_type = "one_button_page",
  record_audio_method = "aws_pyin",
  answer_meta_data = NULL,
  get_answer = get_answer_null,
  save_answer = FALSE,
  stimuli_reactive = FALSE,
  midi_device = " ",
  show_aws_controls = FALSE,
  page_label = "x",
  button_text = "Next",
  play_button_text = "Play",
  note_length = 0.5,
  sound = "piano",
  asChord = FALSE,
  ascending = TRUE,
  start_note = 0,
  end_note = "end",
  dur_list = "null",
  show_record_button = FALSE,
  auto_next_page = FALSE,
  choices = NULL,
  user_rating = FALSE,
  page_text_first = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Present Stimuli
}
