% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_music.R
\name{present_stimuli_midi_notes_both}
\alias{present_stimuli_midi_notes_both}
\title{Present midi notes in both visual and auditory modalities}
\usage{
present_stimuli_midi_notes_both(
  stimuli,
  note_length,
  sound = "piano",
  asChord = FALSE,
  play_button_text = "Play",
  ascending = TRUE,
  visual_music_notation_id = "sheet_music",
  play_button_id = "playButton",
  button_area_id = "button_area",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Present midi notes in both visual and auditory modalities
}
