% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_music.R
\name{present_stimuli_midi_notes_both}
\alias{present_stimuli_midi_notes_both}
\title{Present midi notes in both visual and auditory modalities}
\usage{
present_stimuli_midi_notes_both(
  stimuli,
  note_length = 0.5,
  sound = "piano",
  asChord = FALSE,
  play_button_text = "Play",
  ascending = TRUE,
  visual_music_notation_id = "sheet_music",
  play_button_id = "playButton",
  button_area_id = "button_area",
  sheet_music_start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  sheet_music_id = "sheet_music",
  page_type = "null",
  durations = NULL,
  trigger_start_of_stimulus_fun = wrap_js_fun_body("console.log('Stimulus started!');"),
  trigger_end_of_stimulus_fun = wrap_js_fun_body("console.log('Stimulus finished!');"),
  clef = "auto",
  first_note_message = psychTestR::i18n("first_note_is"),
  transposed_message = psychTestR::i18n("transposed"),
  play_first_note_button_text = psychTestR::i18n("play_first_note"),
  key = NULL
)
}
\arguments{
\item{key}{}
}
\description{
Present midi notes in both visual and auditory modalities
}
