% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_stimuli_music.R
\name{present_stimuli_midi_notes_visual}
\alias{present_stimuli_midi_notes_visual}
\title{Present MIDI notes as musical notation}
\usage{
present_stimuli_midi_notes_visual(
  stimuli,
  note_length,
  asChord = FALSE,
  ascending,
  id = "sheet_music",
  present_div = TRUE,
  clef = "auto",
  transpose_visual_notation = 0L,
  audio_play_button_id = "playButton",
  sheet_music_start_hidden = FALSE,
  durations = NULL
)
}
\arguments{
\item{durations}{}
}
\description{
Present MIDI notes as musical notation
}
