% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psych_test_snap}
\alias{psych_test_snap}
\title{Useful function for randomising which of 2 blocks goes first (or forcing it with force_snap)}
\usage{
psych_test_snap(
  block1,
  block2,
  condition1_name = "1",
  condition2_name = "2",
  force_snap = NULL
)
}
\arguments{
\item{force_snap}{}
}
\description{
Useful function for randomising which of 2 blocks goes first (or forcing it with force_snap)
}
