% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{record_audio_block}
\alias{record_audio_block}
\title{A block of record audio pages}
\usage{
record_audio_block(
  no_pages,
  feedback = NULL,
  get_answer = musicassessr::get_answer_pyin,
  page_title = "Record audio",
  page_text = "Click record to record some audio",
  copy_audio_to = NULL
)
}
\arguments{
\item{page_title}{}
}
\value{

}
\description{
A block of record audio pages
}
