% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{record_audio_block}
\alias{record_audio_block}
\title{A block of record audio pages}
\usage{
record_audio_block(
  no_pages,
  label = "record_audio",
  feedback = NULL,
  get_answer = get_answer_pyin,
  page_title = psychTestR::i18n("Record_audio"),
  page_text = psychTestR::i18n("click_to_record_audio"),
  db_vars = NULL
)
}
\arguments{
\item{db_vars}{}
}
\description{
A block of record audio pages
}
