% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_audio_page.R
\name{record_audio_page}
\alias{record_audio_page}
\title{Page to record audio in psychTestR}
\usage{
record_audio_page(
  body = " ",
  label = "record_audio_page",
  stimuli = " ",
  stimuli_reactive = FALSE,
  page_text = " ",
  page_title = " ",
  interactive = FALSE,
  show_record_button = TRUE,
  get_answer = get_answer_save_aws_key,
  answer_meta_data = tibble::tibble(),
  show_aws_controls = FALSE,
  button_text = psychTestR::i18n("Record"),
  stop_button_text = psychTestR::i18n("Stop"),
  record_duration = NULL,
  on_complete = NULL,
  auto_next_page = FALSE,
  save_answer = TRUE,
  page_text_first = TRUE,
  happy_with_response = FALSE,
  attempts_left = NULL,
  max_goes_forced = FALSE,
  max_goes = 1,
  show_progress = FALSE,
  melody_no = 0,
  total_no_melodies = 0,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Page to record audio in psychTestR
}
