% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_midi_page.R
\name{record_midi_page}
\alias{record_midi_page}
\title{Page to record MIDI in psychTestR}
\usage{
record_midi_page(
  body = " ",
  label = "record_midi_page",
  stimuli = " ",
  stimuli_reactive = FALSE,
  page_text = " ",
  page_title = " ",
  interactive = FALSE,
  show_record_button = TRUE,
  get_answer = get_answer_midi,
  answer_meta_data = tibble::tibble(),
  show_aws_controls,
  button_text = "Record",
  stop_button_text = "Stop",
  record_duration = NULL,
  on_complete = NULL,
  auto_next_page = FALSE,
  save_answer = TRUE,
  page_text_first = TRUE,
  happy_with_response = FALSE,
  attempts_left = NULL,
  max_goes_forced = FALSE,
  autoInstantiate = FALSE,
  midi_device,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Page to record MIDI in psychTestR
}
