% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_midi_page.R
\name{record_midi_page}
\alias{record_midi_page}
\title{Record MIDI page}
\usage{
record_midi_page(
  body = " ",
  label = "record_midi_page",
  stimuli = " ",
  stimuli_reactive = FALSE,
  page_text = " ",
  page_title = " ",
  interactive = FALSE,
  note_no = "max",
  show_record_button = TRUE,
  get_answer = get_answer_midi,
  answer_meta_data = 0,
  autoInstantiate = FALSE,
  midi_device,
  button_text = "Record",
  auto_next_page = FALSE,
  page_text_first = TRUE,
  happy_with_response = FALSE,
  attempts_left = NULL,
  stop_button_text = "Stop",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Record MIDI page
}
