% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{rhythmic_melody_trials}
\alias{rhythmic_melody_trials}
\title{Rhythmic melody trials block}
\usage{
rhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = NULL,
  feedback = FALSE,
  get_answer = musicassessr::get_answer_pyin,
  sound = "piano",
  page_text = psychTestR::i18n("sing_melody_trial"),
  page_title = "Sing This Melody Plus Rhythm",
 
    instruction_text = "Now you will hear melodies with rhythms. Please try and sing the melodies with the correct rhythm."
)
}
\arguments{
\item{instruction_text}{}
}
\value{

}
\description{
Rhythmic melody trials block
}
