% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{rhythmic_melody_trials}
\alias{rhythmic_melody_trials}
\title{Rhythmic melody trials block}
\usage{
rhythmic_melody_trials(
  var_name = "rhythmic_melody",
  module_name = "rhythmic_melodies",
  page_text = psychTestR::i18n("rhythmic_melody_trial_page_text"),
  page_title = psychTestR::i18n("rhythmic_melody_trial_page_title"),
  instruction_text = if (melody_block_paradigm == "sing_melody_first")
    psychTestR::i18n("rhythmic_melody_trial_instruction_text_sing_melody_first") else
    psychTestR::i18n("rhythmic_melody_trial_instruction_text"),
  sampler_function = sample_rhythmic,
  item_bank,
  num_items = integer(),
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production,
  sound = "piano",
  page_type = "record_audio_page",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_progress = TRUE,
  sheet_music_start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  presampled = FALSE,
  get_similarity_to_previous_melody = FALSE,
  volume_meter = FALSE,
  volume_meter_type = "default",
  melody_block_paradigm = c("standard", "sing_melody_first",
    "learn_phase_visual_display_modality"),
  singing_trials = TRUE,
  review = FALSE,
  phase = c("test", "learn", "review"),
  first_note_message = psychTestR::i18n("first_note_is"),
  transposed_message = "Please note, this is transposed for your instrument.",
  play_first_note_button_text = psychTestR::i18n("play_first_note"),
  learn_test_paradigm = FALSE,
  sample_item_bank_via_api = FALSE,
  start_from_sampled_trial_no = 1L,
  pass_items_through_url_parameter = FALSE,
  asynchronous_api_mode = FALSE,
  mute_midi_playback = FALSE
)
}
\arguments{
\item{mute_midi_playback}{}
}
\description{
Rhythmic melody trials block
}
