% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{rhythmic_melody_trials}
\alias{rhythmic_melody_trials}
\title{Rhythmic melody trials block}
\usage{
rhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production,
  sound = "piano",
  page_text = "Press play to hear the melody. Play back the melody plus rhythm.",
  page_title = "Play This Melody Plus Rhythm",
  page_type = "record_audio_page",
 
    instruction_text = "Now you will hear melodies with rhythms. Please try and play the melodies with the correct rhythm.",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = musicassessr::rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  display_modality = "auditory",
  show_record_button = FALSE,
  show_progress = TRUE,
  module_name = "rhythmic_melodies",
  start_hidden = FALSE,
  sound_only_first_melody_note = FALSE,
  show_sheet_music = FALSE,
  sheet_music_id = "sheet_music",
  give_first_melody_note = FALSE,
  sampler_function = sample_rhythmic,
  presampled = FALSE
)
}
\arguments{
\item{presampled}{}
}
\value{

}
\description{
Rhythmic melody trials block
}
