% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{rhythmic_melody_trials}
\alias{rhythmic_melody_trials}
\title{Rhythmic melody trials block}
\usage{
rhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = NULL,
  feedback = FALSE,
  page_title = "Sing This Melody Plus Rhythm",
  get_answer = musicassessr::get_answer_save_aws_key
)
}
\arguments{
\item{get_answer}{}
}
\value{

}
\description{
Rhythmic melody trials block
}
