% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_item_bank.R
\name{sample_rhythmic}
\alias{sample_rhythmic}
\title{Sample rhythmic}
\usage{
sample_rhythmic(item_bank, num_items, id = "rhythmic_melody", phase = NULL)
}
\arguments{
\item{phase}{Currently not used by required.}
}
\description{
Sample rhythmic
}
