% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_melodic_production.R
\name{score_melodic_production}
\alias{score_melodic_production}
\title{Score using melodic production measures}
\usage{
score_melodic_production(
  user_melody_freq = numeric(),
  user_melody_input = numeric(),
  user_duration_input = numeric(),
  user_onset_input = numeric(),
  stimuli = numeric(),
  stimuli_durations = numeric(),
  pyin_pitch_track = tibble::tibble(),
  user_response_midi_note_off = numeric(),
  onsets_noteoff = numeric(),
  answer_meta_data = tibble::tibble(),
  as_tb = FALSE,
  additional_scoring_measures = NULL,
  with_pmi = FALSE,
  compute_accuracy_measures_aligned = TRUE
)
}
\arguments{
\item{user_melody_freq}{Can be NULL for MIDI.}

\item{compute_accuracy_measures_aligned}{}
}
\description{
Score using melodic production measures
}
