% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_custom.R
\name{select_musical_instrument_page}
\alias{select_musical_instrument_page}
\title{Page to select what instrument the participant is using}
\usage{
select_musical_instrument_page(
  set_range_based_on_selection = TRUE,
  include_other_in_dropdown = FALSE,
  with_voice = FALSE,
  prompt = psychTestR::i18n("instrument_selection_message")
)
}
\arguments{
\item{set_range_based_on_selection}{Should the range be set based on the selection?}

\item{include_other_in_dropdown}{Include Other as an option?}

\item{with_voice}{Include Voice as an option?}

\item{prompt}{Prompt?}
}
\description{
Page to select what instrument the participant is using
}
