% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_pages}
\alias{setup_pages}
\title{Setup pages for musicassessr test}
\usage{
setup_pages(
  input_type = c("microphone", "midi_keyboard", "midi_keyboard_and_microphone",
    "midi_keyboard_or_microphone", "key_presses"),
  headphones = TRUE,
  SNR_test = TRUE,
  min_SNR = 14,
  get_user_info = TRUE,
  demo = FALSE,
  get_instrument_range = FALSE,
  absolute_url = character(),
  select_instrument = FALSE,
  get_instrument_range_musical_notation = FALSE,
  adjust_range = FALSE,
  test_type = c("voice", "instrument"),
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  report_SNR = FALSE,
  concise_wording = TRUE,
  skip_setup = FALSE,
  get_self_chosen_anonymous_id = FALSE,
  musical_instrument = FALSE,
  allow_SNR_failure = FALSE,
  requirements_page = TRUE,
  playful_volume_meter_setup = FALSE,
  fake_range = FALSE,
  use_musicassessr_db = FALSE,
  show_microphone_type_page = TRUE
)
}
\arguments{
\item{input_type}{The form of musicassessr test.}

\item{headphones}{Deploy a headphone setup page.}

\item{SNR_test}{Deploy an SNR test.}

\item{min_SNR}{The minimum accepted SNR value.}

\item{get_user_info}{Get user info from their browser.}

\item{demo}{Deploy demo test.}

\item{get_instrument_range}{Get the user's pitch/instrument range at test time.}

\item{absolute_url}{If in production, what is the absolute URL.}

\item{select_instrument}{Deploy instrument selection page.}

\item{get_instrument_range_musical_notation}{When displaying instrument range, provide visual music notation.}

\item{adjust_range}{Adjust range.}

\item{test_type}{Voice vs. instrument test.}

\item{microphone_test}{Deploy microphone test page.}

\item{allow_repeat_SNR_tests}{Allow repeated SNR tests, if FALSE, then participant only gets one go and the test will fail if the SNR test fails.}

\item{report_SNR}{Should the SNR be reported to the user.}

\item{concise_wording}{Whether the wording should be concise or not.}

\item{skip_setup}{Whether to skip setup. Can be TRUE (skip whole setup), FALSE or "except_microphone" (only setup the microphone but no other steps).}

\item{get_self_chosen_anonymous_id}{Whether to ask participant to provide an anonymous ID.}

\item{musical_instrument}{Whether the participant is required to have a musical instrument.}

\item{allow_SNR_failure}{If TRUE, allow user to continue even if they fail the SNR test.}

\item{requirements_page}{Show a requirements page?}

\item{playful_volume_meter_setup}{Should there be some additional functionality to demo the playful volume meter?}

\item{fake_range}{Should the instrument/voice range be faked with a default?}

\item{use_musicassessr_db}{Is musicassessr_db being used?}

\item{show_microphone_type_page}{Should you ask the user what kind of microphone they are using?}
}
\description{
Setup pages for musicassessr test
}
