% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_pages}
\alias{setup_pages}
\title{Setup pages for musicassessr test}
\usage{
setup_pages(
  input = c("microphone", "midi_keyboard", "midi_keyboard_and_microphone",
    "midi_keyboard_or_microphone"),
  headphones = TRUE,
  SNR_test = TRUE,
  min_SNR = 14,
  get_user_info = TRUE,
  demo = FALSE,
  get_instrument_range = FALSE,
  absolute_url,
  select_instrument = FALSE,
  get_instrument_range_musical_notation = FALSE,
  adjust_range = FALSE,
  test_type = c("voice", "instrument"),
  microphone_test = TRUE
)
}
\arguments{
\item{microphone_test}{}
}
\value{

}
\description{
Setup pages for musicassessr test
}
