% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{share_score_page}
\alias{share_score_page}
\title{A page for sharing scores.}
\usage{
share_score_page(
  test_name,
  url,
  hashtag = "CanISing",
  socials = TRUE,
  leaderboard_name = "leaderboard.rda",
  distribution_mean = NULL,
  distribution_sd = NULL,
  data_not_from_leaderboard = NULL
)
}
\arguments{
\item{data_not_from_leaderboard}{}
}
\value{

}
\description{
A page for sharing scores.
}
