% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{share_score_page}
\alias{share_score_page}
\title{A page for sharing scores.}
\usage{
share_score_page(test_name, url, hashtag = "CanISing", socials = TRUE)
}
\arguments{
\item{hashtag}{}
}
\value{

}
\description{
A page for sharing scores.
}
