% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_range.R
\name{sing_happy_birthday_page}
\alias{sing_happy_birthday_page}
\title{A page to identify a user's singing range by asking them to sing Happy Birthday}
\usage{
sing_happy_birthday_page(
  feedback = FALSE,
  label = "sing_hbd",
  text = psychTestR::i18n("sing_hbd")
)
}
\arguments{
\item{text}{}
}
\description{
A page to identify a user's singing range by asking them to sing Happy Birthday
}
