% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_blocks.R
\name{sing_rhythmic_melody_trials}
\alias{sing_rhythmic_melody_trials}
\title{Sing rhythmic melody trial block}
\usage{
sing_rhythmic_melody_trials(
  item_bank,
  num_items,
  num_examples = 0L,
  feedback = FALSE,
  get_answer = get_answer_pyin_melodic_production,
  sound = "piano",
  page_text = psychTestR::i18n("sing_melody_trial"),
  page_title = "Sing This Melody Plus Rhythm",
  page_type = "record_audio_page",
 
    instruction_text = "Now you will hear melodies with rhythms. Please try and sing the melodies with the correct rhythm.",
  get_trial_characteristics_function = NULL,
  item_characteristics_sampler_function = NULL,
  item_characteristics_pars = NULL,
  rel_to_abs_mel_function = musicassessr::rel_to_abs_mel_mean_centred,
  max_goes = 3L,
  max_goes_forced = FALSE,
  give_first_melody_note = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{show_progress}{}
}
\value{

}
\description{
Sing rhythmic melody trial block
}
