% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{test_headphones_page}
\alias{test_headphones_page}
\title{Test headphones page}
\usage{
test_headphones_page(concise_wording = FALSE)
}
\arguments{
\item{concise_wording}{}
}
\value{

}
\description{
Test headphones page
}
