% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{test_headphones_page}
\alias{test_headphones_page}
\title{test_headphones_page
Essentially a wrapper for volume_calibration_page with a default audio file.}
\usage{
test_headphones_page()
}
\value{

}
\description{
test_headphones_page
Essentially a wrapper for volume_calibration_page with a default audio file.
}
