% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{tidy_melodies}
\alias{tidy_melodies}
\title{Tidy melodies from psychTestR results object}
\usage{
tidy_melodies(melody_results)
}
\arguments{
\item{melody_results}{}
}
\value{

}
\description{
Tidy melodies from psychTestR results object
}
