% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{wrap_musicassessr_timeline}
\alias{wrap_musicassessr_timeline}
\title{Wrap a timeline with musicassessr dict}
\usage{
wrap_musicassessr_timeline(tl, language = "en")
}
\arguments{
\item{language}{}
}
\description{
Wrap a timeline with musicassessr dict
}
